+++ STM8S_lib.cscope_file_list
@@ -1,0 +1,52 @@
+"K:\projets\github\stm8s-sdcc\inc\stm8s_uart2.h"
+"K:\projets\github\stm8s-sdcc\inc\stm8s_tim3.h"
+"K:\projets\github\stm8s-sdcc\src\stm8s_tim1.c"
+"K:\projets\github\stm8s-sdcc\src\stm8s_flash.c"
+"K:\projets\github\stm8s-sdcc\inc\stm8s.h"
+"K:\projets\github\stm8s-sdcc\src\stm8s_uart1.c"
+"K:\projets\github\stm8s-sdcc\src\stm8s_tim4.c"
+"K:\projets\github\stm8s-sdcc\inc\stm8s_spi.h"
+"K:\projets\github\stm8s-sdcc\src\stm8s_tim2.c"
+"K:\projets\github\stm8s-sdcc\inc\stm8s_i2c.h"
+"K:\projets\github\stm8s-sdcc\src\stm8s_can.c"
+"K:\projets\github\stm8s-sdcc\inc\stm8s_flash.h"
+"K:\projets\github\stm8s-sdcc\src\stm8s_iwdg.c"
+"K:\projets\github\stm8s-sdcc\inc\stm8s_beep.h"
+"K:\projets\github\stm8s-sdcc\src\stm8s_itc.c"
+"K:\projets\github\stm8s-sdcc\src\stm8s_wwdg.c"
+"K:\projets\github\stm8s-sdcc\inc\stm8s_can.h"
+"K:\projets\github\stm8s-sdcc\src\stm8s_awu.c"
+"K:\projets\github\stm8s-sdcc\inc\stm8s_uart3.h"
+"K:\projets\github\stm8s-sdcc\inc\stm8s_adc1.h"
+"K:\projets\github\stm8s-sdcc\inc\stm8s_uart1.h"
+"K:\projets\github\stm8s-sdcc\inc\stm8s_uart4.h"
+"K:\projets\github\stm8s-sdcc\inc\stm8s_tim4.h"
+"K:\projets\github\stm8s-sdcc\src\stm8s_beep.c"
+"K:\projets\github\stm8s-sdcc\inc\stm8s_iwdg.h"
+"K:\projets\github\stm8s-sdcc\src\stm8s_adc1.c"
+"K:\projets\github\stm8s-sdcc\src\stm8s_clk_HSIPrescalerConfig.c"
+"K:\projets\github\stm8s-sdcc\inc\stm8s_rst.h"
+"K:\projets\github\stm8s-sdcc\inc\stm8s_wwdg.h"
+"K:\projets\github\stm8s-sdcc\inc\stm8s_tim2.h"
+"K:\projets\github\stm8s-sdcc\src\stm8s_uart4.c"
+"K:\projets\github\stm8s-sdcc\src\stm8s_gpio.c"
+"K:\projets\github\stm8s-sdcc\inc\stm8s_adc2.h"
+"K:\projets\github\stm8s-sdcc\src\stm8s_tim5.c"
+"K:\projets\github\stm8s-sdcc\inc\stm8s_tim1.h"
+"K:\projets\github\stm8s-sdcc\src\stm8s_tim3.c"
+"K:\projets\github\stm8s-sdcc\inc\stm8s_itc.h"
+"K:\projets\github\stm8s-sdcc\src\stm8s_spi.c"
+"K:\projets\github\stm8s-sdcc\inc\stm8s_gpio.h"
+"K:\projets\github\stm8s-sdcc\src\stm8s_rst.c"
+"K:\projets\github\stm8s-sdcc\src\stm8s_tim6.c"
+"K:\projets\github\stm8s-sdcc\inc\stm8s_exti.h"
+"K:\projets\github\stm8s-sdcc\inc\stm8s_clk.h"
+"K:\projets\github\stm8s-sdcc\src\stm8s_i2c.c"
+"K:\projets\github\stm8s-sdcc\src\stm8s_clk.c"
+"K:\projets\github\stm8s-sdcc\inc\stm8s_awu.h"
+"K:\projets\github\stm8s-sdcc\inc\stm8s_tim6.h"
+"K:\projets\github\stm8s-sdcc\src\stm8s_uart2.c"
+"K:\projets\github\stm8s-sdcc\src\stm8s_exti.c"
+"K:\projets\github\stm8s-sdcc\inc\stm8s_tim5.h"
+"K:\projets\github\stm8s-sdcc\src\stm8s_uart3.c"
+"K:\projets\github\stm8s-sdcc\src\stm8s_adc2.c"
