+++ src/stm8s_clk_GetClockFreq.c
@@ -1,0 +1,101 @@
+/**
+  ******************************************************************************
+  * @file    stm8s_clk.c
+  * @author  MCD Application Team
+  * @version V2.2.0
+  * @date    30-September-2014
+  * @brief   This file contains all the functions for the CLK peripheral.
+   ******************************************************************************
+  * @attention
+  *
+  * <h2><center>&copy; COPYRIGHT 2014 STMicroelectronics</center></h2>
+  *
+  * Licensed under MCD-ST Liberty SW License Agreement V2, (the "License");
+  * You may not use this file except in compliance with the License.
+  * You may obtain a copy of the License at:
+  *
+  *        http://www.st.com/software_license_agreement_liberty_v2
+  *
+  * Unless required by applicable law or agreed to in writing, software
+  * distributed under the License is distributed on an "AS IS" BASIS,
+  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  * See the License for the specific language governing permissions and
+  * limitations under the License.
+  *
+  ******************************************************************************
+  */
+
+/* Includes ------------------------------------------------------------------*/
+
+#include "stm8s_clk.h"
+
+/** @addtogroup STM8S_StdPeriph_Driver
+  * @{
+  */
+/* Private typedef -----------------------------------------------------------*/
+/* Private define ------------------------------------------------------------*/
+/* Private macro -------------------------------------------------------------*/
+/* Private variables ---------------------------------------------------------*/
+/* Private function prototypes -----------------------------------------------*/
+
+/* Private Constants ---------------------------------------------------------*/
+
+/**
+  * @addtogroup CLK_Private_Constants
+  * @{
+  */
+
+static CONST uint8_t HSIDivExp[4] = {0, 1, 2, 3}; /*!< Holds the HSI Divider exponents */
+
+/**
+  * @}
+  */
+
+/* Public functions ----------------------------------------------------------*/
+/**
+  * @addtogroup CLK_Public_Functions
+  * @{
+  */
+
+/**
+  * @brief  This function returns the frequencies of different on chip clocks.
+  * @param  None
+  * @retval the master clock frequency
+  */
+uint32_t CLK_GetClockFreq(void)
+{
+  uint32_t clockfrequency = 0;
+  CLK_Source_TypeDef clocksource = CLK_SOURCE_HSI;
+  uint8_t tmp = 0;
+
+  /* Get CLK source. */
+  clocksource = (CLK_Source_TypeDef)CLK->CMSR;
+
+  if (clocksource == CLK_SOURCE_HSI)
+  {
+    tmp = (uint8_t)(CLK->CKDIVR & CLK_CKDIVR_HSIDIV);
+    tmp = (uint8_t)(tmp >> 3);
+    clockfrequency = HSI_VALUE >> HSIDivExp[tmp];
+  }
+  else if ( clocksource == CLK_SOURCE_LSI)
+  {
+    clockfrequency = LSI_VALUE;
+  }
+  else
+  {
+    clockfrequency = HSE_VALUE;
+  }
+
+  return((uint32_t)clockfrequency);
+}
+
+/**
+  * @}
+  */
+
+/**
+  * @}
+  */
+
+
+/************************ (C) COPYRIGHT STMicroelectronics *****END OF FILE****/
