+++ src/stm8s_clk_HSIPrescalerConfig.c
@@ -1,0 +1,70 @@
+/**
+  ******************************************************************************
+  * @file    stm8s_clk_HSIPrescalerConfig.c
+  * @author  MCD Application Team
+  * @version V2.2.0
+  * @date    30-September-2014
+  * @brief   This file contains all the functions for the CLK peripheral.
+   ******************************************************************************
+  * @attention
+  *
+  * <h2><center>&copy; COPYRIGHT 2014 STMicroelectronics</center></h2>
+  *
+  * Licensed under MCD-ST Liberty SW License Agreement V2, (the "License");
+  * You may not use this file except in compliance with the License.
+  * You may obtain a copy of the License at:
+  *
+  *        http://www.st.com/software_license_agreement_liberty_v2
+  *
+  * Unless required by applicable law or agreed to in writing, software
+  * distributed under the License is distributed on an "AS IS" BASIS,
+  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  * See the License for the specific language governing permissions and
+  * limitations under the License.
+  *
+  ******************************************************************************
+  */
+
+/* Includes ------------------------------------------------------------------*/
+
+#include "stm8s_clk.h"
+
+/** @addtogroup STM8S_StdPeriph_Driver
+  * @{
+  */
+/* Private typedef -----------------------------------------------------------*/
+/* Private define ------------------------------------------------------------*/
+/* Private macro -------------------------------------------------------------*/
+/* Private variables ---------------------------------------------------------*/
+/* Private function prototypes -----------------------------------------------*/
+
+/* Private Constants ---------------------------------------------------------*/
+
+/**
+  * @brief  Configures the HSI clock dividers.
+  * @param   HSIPrescaler : Specifies the HSI clock divider to apply.
+  * This parameter can be any of the @ref CLK_Prescaler_TypeDef enumeration.
+  * @retval None
+  */
+void CLK_HSIPrescalerConfig(CLK_Prescaler_TypeDef HSIPrescaler)
+{
+  /* check the parameters */
+  assert_param(IS_CLK_HSIPRESCALER_OK(HSIPrescaler));
+
+  /* Clear High speed internal clock prescaler */
+  CLK->CKDIVR &= (uint8_t)(~CLK_CKDIVR_HSIDIV);
+
+  /* Set High speed internal clock prescaler */
+  CLK->CKDIVR |= (uint8_t)HSIPrescaler;
+}
+
+/**
+  * @}
+  */
+
+/**
+  * @}
+  */
+
+
+/************************ (C) COPYRIGHT STMicroelectronics *****END OF FILE****/
