+++ src/stm8s_gpio.c
@@ -16,8 +16,8 @@
   *
   *        http://www.st.com/software_license_agreement_liberty_v2
   *
-  * Unless required by applicable law or agreed to in writing, software 
-  * distributed under the License is distributed on an "AS IS" BASIS, 
+  * Unless required by applicable law or agreed to in writing, software
+  * distributed under the License is distributed on an "AS IS" BASIS,
   * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
   * See the License for the specific language governing permissions and
   * limitations under the License.
@@ -59,78 +59,6 @@ void GPIO_DeInit(GPIO_TypeDef* GPIOx)
 }
 
 /**
-  * @brief  Initializes the GPIOx according to the specified parameters.
-  * @param  GPIOx : Select the GPIO peripheral number (x = A to I).
-  * @param  GPIO_Pin : This parameter contains the pin number, it can be any value
-  *         of the @ref GPIO_Pin_TypeDef enumeration.
-  * @param  GPIO_Mode : This parameter can be a value of the
-  *         @Ref GPIO_Mode_TypeDef enumeration.
-  * @retval None
-  */
-
-void GPIO_Init(GPIO_TypeDef* GPIOx, GPIO_Pin_TypeDef GPIO_Pin, GPIO_Mode_TypeDef GPIO_Mode)
-{
-  /*----------------------*/
-  /* Check the parameters */
-  /*----------------------*/
-  
-  assert_param(IS_GPIO_MODE_OK(GPIO_Mode));
-  assert_param(IS_GPIO_PIN_OK(GPIO_Pin));
-  
-  /* Reset corresponding bit to GPIO_Pin in CR2 register */
-  GPIOx->CR2 &= (uint8_t)(~(GPIO_Pin));
-  
-  /*-----------------------------*/
-  /* Input/Output mode selection */
-  /*-----------------------------*/
-  
-  if ((((uint8_t)(GPIO_Mode)) & (uint8_t)0x80) != (uint8_t)0x00) /* Output mode */
-  {
-    if ((((uint8_t)(GPIO_Mode)) & (uint8_t)0x10) != (uint8_t)0x00) /* High level */
-    {
-      GPIOx->ODR |= (uint8_t)GPIO_Pin;
-    } 
-    else /* Low level */
-    {
-      GPIOx->ODR &= (uint8_t)(~(GPIO_Pin));
-    }
-    /* Set Output mode */
-    GPIOx->DDR |= (uint8_t)GPIO_Pin;
-  } 
-  else /* Input mode */
-  {
-    /* Set Input mode */
-    GPIOx->DDR &= (uint8_t)(~(GPIO_Pin));
-  }
-  
-  /*------------------------------------------------------------------------*/
-  /* Pull-Up/Float (Input) or Push-Pull/Open-Drain (Output) modes selection */
-  /*------------------------------------------------------------------------*/
-  
-  if ((((uint8_t)(GPIO_Mode)) & (uint8_t)0x40) != (uint8_t)0x00) /* Pull-Up or Push-Pull */
-  {
-    GPIOx->CR1 |= (uint8_t)GPIO_Pin;
-  } 
-  else /* Float or Open-Drain */
-  {
-    GPIOx->CR1 &= (uint8_t)(~(GPIO_Pin));
-  }
-  
-  /*-----------------------------------------------------*/
-  /* Interrupt (Input) or Slope (Output) modes selection */
-  /*-----------------------------------------------------*/
-  
-  if ((((uint8_t)(GPIO_Mode)) & (uint8_t)0x20) != (uint8_t)0x00) /* Interrupt or Slow slope */
-  {
-    GPIOx->CR2 |= (uint8_t)GPIO_Pin;
-  } 
-  else /* No external interrupt or No slope control */
-  {
-    GPIOx->CR2 &= (uint8_t)(~(GPIO_Pin));
-  }
-}
-
-/**
   * @brief  Writes data to the specified GPIO data port.
   * @note   The port must be configured in output mode.
   * @param  GPIOx : Select the GPIO peripheral number (x = A to I).
@@ -141,32 +69,6 @@ void GPIO_Init(GPIO_TypeDef* GPIOx, GPIO_Pin_TypeDef GPIO_Pin, GPIO_Mode_TypeDef
 void GPIO_Write(GPIO_TypeDef* GPIOx, uint8_t PortVal)
 {
   GPIOx->ODR = PortVal;
-}
-
-/**
-  * @brief  Writes high level to the specified GPIO pins.
-  * @note   The port must be configured in output mode.  
-  * @param  GPIOx : Select the GPIO peripheral number (x = A to I).
-  * @param  PortPins : Specifies the pins to be turned high to the port output.
-  *         data register.
-  * @retval None
-  */
-void GPIO_WriteHigh(GPIO_TypeDef* GPIOx, GPIO_Pin_TypeDef PortPins)
-{
-  GPIOx->ODR |= (uint8_t)PortPins;
-}
-
-/**
-  * @brief  Writes low level to the specified GPIO pins.
-  * @note   The port must be configured in output mode.  
-  * @param  GPIOx : Select the GPIO peripheral number (x = A to I).
-  * @param  PortPins : Specifies the pins to be turned low to the port output.
-  *         data register.
-  * @retval None
-  */
-void GPIO_WriteLow(GPIO_TypeDef* GPIOx, GPIO_Pin_TypeDef PortPins)
-{
-  GPIOx->ODR &= (uint8_t)(~PortPins);
 }
 
 /**
@@ -184,24 +86,13 @@ void GPIO_WriteReverse(GPIO_TypeDef* GPIOx, GPIO_Pin_TypeDef PortPins)
 
 /**
   * @brief  Reads the specified GPIO output data port.
-  * @note   The port must be configured in input mode.  
+  * @note   The port must be configured in input mode.
   * @param  GPIOx : Select the GPIO peripheral number (x = A to I).
   * @retval GPIO output data port value.
   */
 uint8_t GPIO_ReadOutputData(GPIO_TypeDef* GPIOx)
 {
   return ((uint8_t)GPIOx->ODR);
-}
-
-/**
-  * @brief  Reads the specified GPIO input data port.
-  * @note   The port must be configured in input mode.   
-  * @param  GPIOx : Select the GPIO peripheral number (x = A to I).
-  * @retval GPIO input data port value.
-  */
-uint8_t GPIO_ReadInputData(GPIO_TypeDef* GPIOx)
-{
-  return ((uint8_t)GPIOx->IDR);
 }
 
 /**
@@ -227,7 +118,7 @@ void GPIO_ExternalPullUpConfig(GPIO_TypeDef* GPIOx, GPIO_Pin_TypeDef GPIO_Pin, F
   /* Check the parameters */
   assert_param(IS_GPIO_PIN_OK(GPIO_Pin));
   assert_param(IS_FUNCTIONALSTATE_OK(NewState));
-  
+
   if (NewState != DISABLE) /* External Pull-Up Set*/
   {
     GPIOx->CR1 |= (uint8_t)GPIO_Pin;
@@ -240,10 +131,10 @@ void GPIO_ExternalPullUpConfig(GPIO_TypeDef* GPIOx, GPIO_Pin_TypeDef GPIO_Pin, F
 /**
   * @}
   */
-  
+
 /**
   * @}
   */
-  
+
 
 /************************ (C) COPYRIGHT STMicroelectronics *****END OF FILE****/
