+++ src/stm8s_tim1.c
@@ -16,8 +16,8 @@
   *
   *        http://www.st.com/software_license_agreement_liberty_v2
   *
-  * Unless required by applicable law or agreed to in writing, software 
-  * distributed under the License is distributed on an "AS IS" BASIS, 
+  * Unless required by applicable law or agreed to in writing, software
+  * distributed under the License is distributed on an "AS IS" BASIS,
   * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
   * See the License for the specific language governing permissions and
   * limitations under the License.
@@ -101,53 +101,21 @@ void TIM1_DeInit(void)
 }
 
 /**
-  * @brief  Initializes the TIM1 Time Base Unit according to the specified parameters.
-  * @param  TIM1_Prescaler specifies the Prescaler value.
-  * @param  TIM1_CounterMode specifies the counter mode  from @ref TIM1_CounterMode_TypeDef .
-  * @param  TIM1_Period specifies the Period value.
-  * @param  TIM1_RepetitionCounter specifies the Repetition counter value
-  * @retval None
-  */
-void TIM1_TimeBaseInit(uint16_t TIM1_Prescaler,
-                       TIM1_CounterMode_TypeDef TIM1_CounterMode,
-                       uint16_t TIM1_Period,
-                       uint8_t TIM1_RepetitionCounter)
-{
-  /* Check parameters */
-  assert_param(IS_TIM1_COUNTER_MODE_OK(TIM1_CounterMode));
-  
-  /* Set the Autoreload value */
-  TIM1->ARRH = (uint8_t)(TIM1_Period >> 8);
-  TIM1->ARRL = (uint8_t)(TIM1_Period);
-  
-  /* Set the Prescaler value */
-  TIM1->PSCRH = (uint8_t)(TIM1_Prescaler >> 8);
-  TIM1->PSCRL = (uint8_t)(TIM1_Prescaler);
-  
-  /* Select the Counter Mode */
-  TIM1->CR1 = (uint8_t)((uint8_t)(TIM1->CR1 & (uint8_t)(~(TIM1_CR1_CMS | TIM1_CR1_DIR)))
-                        | (uint8_t)(TIM1_CounterMode));
-  
-  /* Set the Repetition Counter value */
-  TIM1->RCR = TIM1_RepetitionCounter;
-}
-
-/**
   * @brief  Initializes the TIM1 Channel1 according to the specified parameters.
-  * @param  TIM1_OCMode specifies the Output Compare mode from 
+  * @param  TIM1_OCMode specifies the Output Compare mode from
   *         @ref TIM1_OCMode_TypeDef.
-  * @param  TIM1_OutputState specifies the Output State from 
+  * @param  TIM1_OutputState specifies the Output State from
   *         @ref TIM1_OutputState_TypeDef.
-  * @param  TIM1_OutputNState specifies the Complementary Output State 
+  * @param  TIM1_OutputNState specifies the Complementary Output State
   *         from @ref TIM1_OutputNState_TypeDef.
   * @param  TIM1_Pulse specifies the Pulse width value.
-  * @param  TIM1_OCPolarity specifies the Output Compare Polarity from 
+  * @param  TIM1_OCPolarity specifies the Output Compare Polarity from
   *         @ref TIM1_OCPolarity_TypeDef.
   * @param  TIM1_OCNPolarity specifies the Complementary Output Compare Polarity
   *         from @ref TIM1_OCNPolarity_TypeDef.
-  * @param  TIM1_OCIdleState specifies the Output Compare Idle State from 
+  * @param  TIM1_OCIdleState specifies the Output Compare Idle State from
   *         @ref TIM1_OCIdleState_TypeDef.
-  * @param  TIM1_OCNIdleState specifies the Complementary Output Compare Idle 
+  * @param  TIM1_OCNIdleState specifies the Complementary Output Compare Idle
   *         State from @ref TIM1_OCIdleState_TypeDef.
   * @retval None
   */
@@ -168,10 +136,10 @@ void TIM1_OC1Init(TIM1_OCMode_TypeDef TIM1_OCMode,
   assert_param(IS_TIM1_OCN_POLARITY_OK(TIM1_OCNPolarity));
   assert_param(IS_TIM1_OCIDLE_STATE_OK(TIM1_OCIdleState));
   assert_param(IS_TIM1_OCNIDLE_STATE_OK(TIM1_OCNIdleState));
-  
-  /* Disable the Channel 1: Reset the CCE Bit, Set the Output State , 
+
+  /* Disable the Channel 1: Reset the CCE Bit, Set the Output State ,
   the Output N State, the Output Polarity & the Output N Polarity*/
-  TIM1->CCER1 &= (uint8_t)(~( TIM1_CCER1_CC1E | TIM1_CCER1_CC1NE 
+  TIM1->CCER1 &= (uint8_t)(~( TIM1_CCER1_CC1E | TIM1_CCER1_CC1NE
                              | TIM1_CCER1_CC1P | TIM1_CCER1_CC1NP));
   /* Set the Output State & Set the Output N State & Set the Output Polarity &
   Set the Output N Polarity */
@@ -179,17 +147,17 @@ void TIM1_OC1Init(TIM1_OCMode_TypeDef TIM1_OCMode,
                                      | (uint8_t)(TIM1_OutputNState & TIM1_CCER1_CC1NE))
                            | (uint8_t)( (uint8_t)(TIM1_OCPolarity  & TIM1_CCER1_CC1P)
                                        | (uint8_t)(TIM1_OCNPolarity & TIM1_CCER1_CC1NP)));
-  
+
   /* Reset the Output Compare Bits & Set the Output Compare Mode */
-  TIM1->CCMR1 = (uint8_t)((uint8_t)(TIM1->CCMR1 & (uint8_t)(~TIM1_CCMR_OCM)) | 
+  TIM1->CCMR1 = (uint8_t)((uint8_t)(TIM1->CCMR1 & (uint8_t)(~TIM1_CCMR_OCM)) |
                           (uint8_t)TIM1_OCMode);
-  
+
   /* Reset the Output Idle state & the Output N Idle state bits */
   TIM1->OISR &= (uint8_t)(~(TIM1_OISR_OIS1 | TIM1_OISR_OIS1N));
   /* Set the Output Idle state & the Output N Idle state configuration */
-  TIM1->OISR |= (uint8_t)((uint8_t)( TIM1_OCIdleState & TIM1_OISR_OIS1 ) | 
+  TIM1->OISR |= (uint8_t)((uint8_t)( TIM1_OCIdleState & TIM1_OISR_OIS1 ) |
                           (uint8_t)( TIM1_OCNIdleState & TIM1_OISR_OIS1N ));
-  
+
   /* Set the Pulse value */
   TIM1->CCR1H = (uint8_t)(TIM1_Pulse >> 8);
   TIM1->CCR1L = (uint8_t)(TIM1_Pulse);
@@ -199,18 +167,18 @@ void TIM1_OC1Init(TIM1_OCMode_TypeDef TIM1_OCMode,
   * @brief  Initializes the TIM1 Channel2 according to the specified parameters.
   * @param  TIM1_OCMode specifies the Output Compare mode from
   *         @ref TIM1_OCMode_TypeDef.
-  * @param  TIM1_OutputState specifies the Output State from 
+  * @param  TIM1_OutputState specifies the Output State from
   *         @ref TIM1_OutputState_TypeDef.
-  * @param  TIM1_OutputNState specifies the Complementary Output State from 
+  * @param  TIM1_OutputNState specifies the Complementary Output State from
   *         @ref TIM1_OutputNState_TypeDef.
   * @param  TIM1_Pulse specifies the Pulse width value.
-  * @param  TIM1_OCPolarity specifies the Output Compare Polarity from 
+  * @param  TIM1_OCPolarity specifies the Output Compare Polarity from
   *         @ref TIM1_OCPolarity_TypeDef.
   * @param  TIM1_OCNPolarity specifies the Complementary Output Compare Polarity
   *         from @ref TIM1_OCNPolarity_TypeDef.
-  * @param  TIM1_OCIdleState specifies the Output Compare Idle State from 
+  * @param  TIM1_OCIdleState specifies the Output Compare Idle State from
   *         @ref TIM1_OCIdleState_TypeDef.
-  * @param  TIM1_OCNIdleState specifies the Complementary Output Compare Idle 
+  * @param  TIM1_OCNIdleState specifies the Complementary Output Compare Idle
   *         State from @ref TIM1_OCIdleState_TypeDef.
   * @retval None
   */
@@ -231,29 +199,29 @@ void TIM1_OC2Init(TIM1_OCMode_TypeDef TIM1_OCMode,
   assert_param(IS_TIM1_OCN_POLARITY_OK(TIM1_OCNPolarity));
   assert_param(IS_TIM1_OCIDLE_STATE_OK(TIM1_OCIdleState));
   assert_param(IS_TIM1_OCNIDLE_STATE_OK(TIM1_OCNIdleState));
-  
-  /* Disable the Channel 1: Reset the CCE Bit, Set the Output State , 
+
+  /* Disable the Channel 1: Reset the CCE Bit, Set the Output State ,
   the Output N State, the Output Polarity & the Output N Polarity*/
-  TIM1->CCER1 &= (uint8_t)(~( TIM1_CCER1_CC2E | TIM1_CCER1_CC2NE | 
+  TIM1->CCER1 &= (uint8_t)(~( TIM1_CCER1_CC2E | TIM1_CCER1_CC2NE |
                              TIM1_CCER1_CC2P | TIM1_CCER1_CC2NP));
-  
+
   /* Set the Output State & Set the Output N State & Set the Output Polarity &
   Set the Output N Polarity */
-  TIM1->CCER1 |= (uint8_t)((uint8_t)((uint8_t)(TIM1_OutputState & TIM1_CCER1_CC2E  ) | 
-                                     (uint8_t)(TIM1_OutputNState & TIM1_CCER1_CC2NE )) | 
-                           (uint8_t)((uint8_t)(TIM1_OCPolarity  & TIM1_CCER1_CC2P  ) | 
+  TIM1->CCER1 |= (uint8_t)((uint8_t)((uint8_t)(TIM1_OutputState & TIM1_CCER1_CC2E  ) |
+                                     (uint8_t)(TIM1_OutputNState & TIM1_CCER1_CC2NE )) |
+                           (uint8_t)((uint8_t)(TIM1_OCPolarity  & TIM1_CCER1_CC2P  ) |
                                      (uint8_t)(TIM1_OCNPolarity & TIM1_CCER1_CC2NP )));
-  
+
   /* Reset the Output Compare Bits & Set the Output Compare Mode */
-  TIM1->CCMR2 = (uint8_t)((uint8_t)(TIM1->CCMR2 & (uint8_t)(~TIM1_CCMR_OCM)) | 
+  TIM1->CCMR2 = (uint8_t)((uint8_t)(TIM1->CCMR2 & (uint8_t)(~TIM1_CCMR_OCM)) |
                           (uint8_t)TIM1_OCMode);
-  
+
   /* Reset the Output Idle state & the Output N Idle state bits */
   TIM1->OISR &= (uint8_t)(~(TIM1_OISR_OIS2 | TIM1_OISR_OIS2N));
   /* Set the Output Idle state & the Output N Idle state configuration */
-  TIM1->OISR |= (uint8_t)((uint8_t)(TIM1_OISR_OIS2 & TIM1_OCIdleState) | 
+  TIM1->OISR |= (uint8_t)((uint8_t)(TIM1_OISR_OIS2 & TIM1_OCIdleState) |
                           (uint8_t)(TIM1_OISR_OIS2N & TIM1_OCNIdleState));
-  
+
   /* Set the Pulse value */
   TIM1->CCR2H = (uint8_t)(TIM1_Pulse >> 8);
   TIM1->CCR2L = (uint8_t)(TIM1_Pulse);
@@ -261,20 +229,20 @@ void TIM1_OC2Init(TIM1_OCMode_TypeDef TIM1_OCMode,
 
 /**
   * @brief  Initializes the TIM1 Channel3 according to the specified parameters.
-  * @param  TIM1_OCMode specifies the Output Compare mode  from 
+  * @param  TIM1_OCMode specifies the Output Compare mode  from
   *         @ref TIM1_OCMode_TypeDef.
-  * @param  TIM1_OutputState specifies the Output State  
+  * @param  TIM1_OutputState specifies the Output State
   *         from @ref TIM1_OutputState_TypeDef.
   * @param  TIM1_OutputNState specifies the Complementary Output State
   *         from @ref TIM1_OutputNState_TypeDef.
   * @param  TIM1_Pulse specifies the Pulse width value.
-  * @param  TIM1_OCPolarity specifies the Output Compare Polarity  from 
+  * @param  TIM1_OCPolarity specifies the Output Compare Polarity  from
   *         @ref TIM1_OCPolarity_TypeDef.
-  * @param  TIM1_OCNPolarity specifies the Complementary  Output Compare 
+  * @param  TIM1_OCNPolarity specifies the Complementary  Output Compare
   *         Polarity from @ref TIM1_OCNPolarity_TypeDef.
   * @param  TIM1_OCIdleState specifies the Output Compare Idle State
   *         from @ref TIM1_OCIdleState_TypeDef.
-  * @param  TIM1_OCNIdleState specifies the Complementary Output Compare 
+  * @param  TIM1_OCNIdleState specifies the Complementary Output Compare
   *         Idle State  from @ref TIM1_OCIdleState_TypeDef.
   * @retval None
   */
@@ -295,28 +263,28 @@ void TIM1_OC3Init(TIM1_OCMode_TypeDef TIM1_OCMode,
   assert_param(IS_TIM1_OCN_POLARITY_OK(TIM1_OCNPolarity));
   assert_param(IS_TIM1_OCIDLE_STATE_OK(TIM1_OCIdleState));
   assert_param(IS_TIM1_OCNIDLE_STATE_OK(TIM1_OCNIdleState));
-  
-  /* Disable the Channel 1: Reset the CCE Bit, Set the Output State , 
+
+  /* Disable the Channel 1: Reset the CCE Bit, Set the Output State ,
   the Output N State, the Output Polarity & the Output N Polarity*/
-  TIM1->CCER2 &= (uint8_t)(~( TIM1_CCER2_CC3E | TIM1_CCER2_CC3NE | 
+  TIM1->CCER2 &= (uint8_t)(~( TIM1_CCER2_CC3E | TIM1_CCER2_CC3NE |
                              TIM1_CCER2_CC3P | TIM1_CCER2_CC3NP));
   /* Set the Output State & Set the Output N State & Set the Output Polarity &
   Set the Output N Polarity */
   TIM1->CCER2 |= (uint8_t)((uint8_t)((uint8_t)(TIM1_OutputState  & TIM1_CCER2_CC3E   ) |
-                                     (uint8_t)(TIM1_OutputNState & TIM1_CCER2_CC3NE  )) | 
-                           (uint8_t)((uint8_t)(TIM1_OCPolarity   & TIM1_CCER2_CC3P   ) | 
+                                     (uint8_t)(TIM1_OutputNState & TIM1_CCER2_CC3NE  )) |
+                           (uint8_t)((uint8_t)(TIM1_OCPolarity   & TIM1_CCER2_CC3P   ) |
                                      (uint8_t)(TIM1_OCNPolarity  & TIM1_CCER2_CC3NP  )));
-  
+
   /* Reset the Output Compare Bits & Set the Output Compare Mode */
-  TIM1->CCMR3 = (uint8_t)((uint8_t)(TIM1->CCMR3 & (uint8_t)(~TIM1_CCMR_OCM)) | 
+  TIM1->CCMR3 = (uint8_t)((uint8_t)(TIM1->CCMR3 & (uint8_t)(~TIM1_CCMR_OCM)) |
                           (uint8_t)TIM1_OCMode);
-  
+
   /* Reset the Output Idle state & the Output N Idle state bits */
   TIM1->OISR &= (uint8_t)(~(TIM1_OISR_OIS3 | TIM1_OISR_OIS3N));
   /* Set the Output Idle state & the Output N Idle state configuration */
-  TIM1->OISR |= (uint8_t)((uint8_t)(TIM1_OISR_OIS3 & TIM1_OCIdleState) | 
+  TIM1->OISR |= (uint8_t)((uint8_t)(TIM1_OISR_OIS3 & TIM1_OCIdleState) |
                           (uint8_t)(TIM1_OISR_OIS3N & TIM1_OCNIdleState));
-  
+
   /* Set the Pulse value */
   TIM1->CCR3H = (uint8_t)(TIM1_Pulse >> 8);
   TIM1->CCR3L = (uint8_t)(TIM1_Pulse);
@@ -324,7 +292,7 @@ void TIM1_OC3Init(TIM1_OCMode_TypeDef TIM1_OCMode,
 
 /**
   * @brief  Initializes the TIM1 Channel4 according to the specified parameters.
-  * @param  TIM1_OCMode specifies the Output Compare mode  from 
+  * @param  TIM1_OCMode specifies the Output Compare mode  from
   *         @ref TIM1_OCMode_TypeDef.
   * @param  TIM1_OutputState specifies the Output State
   *         from @ref TIM1_OutputState_TypeDef.
@@ -346,17 +314,17 @@ void TIM1_OC4Init(TIM1_OCMode_TypeDef TIM1_OCMode,
   assert_param(IS_TIM1_OUTPUT_STATE_OK(TIM1_OutputState));
   assert_param(IS_TIM1_OC_POLARITY_OK(TIM1_OCPolarity));
   assert_param(IS_TIM1_OCIDLE_STATE_OK(TIM1_OCIdleState));
-  
+
   /* Disable the Channel 4: Reset the CCE Bit */
   TIM1->CCER2 &= (uint8_t)(~(TIM1_CCER2_CC4E | TIM1_CCER2_CC4P));
   /* Set the Output State  &  the Output Polarity */
-  TIM1->CCER2 |= (uint8_t)((uint8_t)(TIM1_OutputState & TIM1_CCER2_CC4E ) |  
+  TIM1->CCER2 |= (uint8_t)((uint8_t)(TIM1_OutputState & TIM1_CCER2_CC4E ) |
                            (uint8_t)(TIM1_OCPolarity  & TIM1_CCER2_CC4P ));
-  
+
   /* Reset the Output Compare Bit  and Set the Output Compare Mode */
-  TIM1->CCMR4 = (uint8_t)((uint8_t)(TIM1->CCMR4 & (uint8_t)(~TIM1_CCMR_OCM)) | 
+  TIM1->CCMR4 = (uint8_t)((uint8_t)(TIM1->CCMR4 & (uint8_t)(~TIM1_CCMR_OCM)) |
                           TIM1_OCMode);
-  
+
   /* Set the Output Idle state */
   if (TIM1_OCIdleState != TIM1_OCIDLESTATE_RESET)
   {
@@ -366,7 +334,7 @@ void TIM1_OC4Init(TIM1_OCMode_TypeDef TIM1_OCMode,
   {
     TIM1->OISR &= (uint8_t)(~TIM1_OISR_OIS4);
   }
-  
+
   /* Set the Pulse value */
   TIM1->CCR4H = (uint8_t)(TIM1_Pulse >> 8);
   TIM1->CCR4L = (uint8_t)(TIM1_Pulse);
@@ -379,9 +347,9 @@ void TIM1_OC4Init(TIM1_OCMode_TypeDef TIM1_OCMode,
   * @param  TIM1_LockLevel specifies the lock level from @ref TIM1_LockLevel_TypeDef.
   * @param  TIM1_DeadTime specifies the dead time value.
   * @param  TIM1_Break specifies the Break state @ref TIM1_BreakState_TypeDef.
-  * @param  TIM1_BreakPolarity specifies the Break polarity from 
+  * @param  TIM1_BreakPolarity specifies the Break polarity from
   *         @ref TIM1_BreakPolarity_TypeDef.
-  * @param  TIM1_AutomaticOutput specifies the Automatic Output configuration 
+  * @param  TIM1_AutomaticOutput specifies the Automatic Output configuration
   *         from @ref TIM1_AutomaticOutput_TypeDef.
   * @retval None
   */
@@ -398,22 +366,22 @@ void TIM1_BDTRConfig(TIM1_OSSIState_TypeDef TIM1_OSSIState,
   assert_param(IS_TIM1_BREAK_STATE_OK(TIM1_Break));
   assert_param(IS_TIM1_BREAK_POLARITY_OK(TIM1_BreakPolarity));
   assert_param(IS_TIM1_AUTOMATIC_OUTPUT_STATE_OK(TIM1_AutomaticOutput));
-  
+
   TIM1->DTR = (uint8_t)(TIM1_DeadTime);
   /* Set the Lock level, the Break enable Bit and the Polarity, the OSSI State,
   the dead time value  and the Automatic Output Enable Bit */
-  
-  TIM1->BKR  =  (uint8_t)((uint8_t)(TIM1_OSSIState | (uint8_t)TIM1_LockLevel)  | 
-                          (uint8_t)((uint8_t)(TIM1_Break | (uint8_t)TIM1_BreakPolarity)  | 
+
+  TIM1->BKR  =  (uint8_t)((uint8_t)(TIM1_OSSIState | (uint8_t)TIM1_LockLevel)  |
+                          (uint8_t)((uint8_t)(TIM1_Break | (uint8_t)TIM1_BreakPolarity)  |
                           (uint8_t)TIM1_AutomaticOutput));
 }
 
 /**
   * @brief  Initializes the TIM1 peripheral according to the specified parameters.
   * @param  TIM1_Channel specifies the input capture channel from TIM1_Channel_TypeDef.
-  * @param  TIM1_ICPolarity specifies the Input capture polarity from  
+  * @param  TIM1_ICPolarity specifies the Input capture polarity from
   *         TIM1_ICPolarity_TypeDef .
-  * @param  TIM1_ICSelection specifies the Input capture source selection from 
+  * @param  TIM1_ICSelection specifies the Input capture source selection from
   *         TIM1_ICSelection_TypeDef.
   * @param  TIM1_ICPrescaler specifies the Input capture Prescaler from
   *         TIM1_ICPSC_TypeDef.
@@ -432,7 +400,7 @@ void TIM1_ICInit(TIM1_Channel_TypeDef TIM1_Channel,
   assert_param(IS_TIM1_IC_SELECTION_OK(TIM1_ICSelection));
   assert_param(IS_TIM1_IC_PRESCALER_OK(TIM1_ICPrescaler));
   assert_param(IS_TIM1_IC_FILTER_OK(TIM1_ICFilter));
-  
+
   if (TIM1_Channel == TIM1_CHANNEL_1)
   {
     /* TI1 Configuration */
@@ -472,15 +440,15 @@ void TIM1_ICInit(TIM1_Channel_TypeDef TIM1_Channel,
 }
 
 /**
-  * @brief  Configures the TIM1 peripheral in PWM Input Mode according to the 
+  * @brief  Configures the TIM1 peripheral in PWM Input Mode according to the
   *         specified parameters.
-  * @param  TIM1_Channel specifies the input capture channel from 
+  * @param  TIM1_Channel specifies the input capture channel from
   *         @ref TIM1_Channel_TypeDef.
-  * @param  TIM1_ICPolarity specifies the Input capture polarity from  
+  * @param  TIM1_ICPolarity specifies the Input capture polarity from
   *         @ref TIM1_ICPolarity_TypeDef .
   * @param  TIM1_ICSelection specifies the Input capture source selection  from
   *         @ref TIM1_ICSelection_TypeDef.
-  * @param  TIM1_ICPrescaler specifies the Input capture Prescaler from  
+  * @param  TIM1_ICPrescaler specifies the Input capture Prescaler from
   *         @ref TIM1_ICPSC_TypeDef.
   * @param  TIM1_ICFilter specifies the Input capture filter value.
   * @retval None
@@ -493,13 +461,13 @@ void TIM1_PWMIConfig(TIM1_Channel_TypeDef TIM1_Channel,
 {
   uint8_t icpolarity = TIM1_ICPOLARITY_RISING;
   uint8_t icselection = TIM1_ICSELECTION_DIRECTTI;
-  
+
   /* Check the parameters */
   assert_param(IS_TIM1_PWMI_CHANNEL_OK(TIM1_Channel));
   assert_param(IS_TIM1_IC_POLARITY_OK(TIM1_ICPolarity));
   assert_param(IS_TIM1_IC_SELECTION_OK(TIM1_ICSelection));
   assert_param(IS_TIM1_IC_PRESCALER_OK(TIM1_ICPrescaler));
-  
+
   /* Select the Opposite Input Polarity */
   if (TIM1_ICPolarity != TIM1_ICPOLARITY_FALLING)
   {
@@ -509,7 +477,7 @@ void TIM1_PWMIConfig(TIM1_Channel_TypeDef TIM1_Channel,
   {
     icpolarity = TIM1_ICPOLARITY_RISING;
   }
-  
+
   /* Select the Opposite Input */
   if (TIM1_ICSelection == TIM1_ICSELECTION_DIRECTTI)
   {
@@ -519,19 +487,19 @@ void TIM1_PWMIConfig(TIM1_Channel_TypeDef TIM1_Channel,
   {
     icselection = TIM1_ICSELECTION_DIRECTTI;
   }
-  
+
   if (TIM1_Channel == TIM1_CHANNEL_1)
   {
     /* TI1 Configuration */
     TI1_Config((uint8_t)TIM1_ICPolarity, (uint8_t)TIM1_ICSelection,
                (uint8_t)TIM1_ICFilter);
-    
+
     /* Set the Input Capture Prescaler value */
     TIM1_SetIC1Prescaler(TIM1_ICPrescaler);
-    
+
     /* TI2 Configuration */
     TI2_Config(icpolarity, icselection, TIM1_ICFilter);
-    
+
     /* Set the Input Capture Prescaler value */
     TIM1_SetIC2Prescaler(TIM1_ICPrescaler);
   }
@@ -540,37 +508,15 @@ void TIM1_PWMIConfig(TIM1_Channel_TypeDef TIM1_Channel,
     /* TI2 Configuration */
     TI2_Config((uint8_t)TIM1_ICPolarity, (uint8_t)TIM1_ICSelection,
                (uint8_t)TIM1_ICFilter);
-    
+
     /* Set the Input Capture Prescaler value */
     TIM1_SetIC2Prescaler(TIM1_ICPrescaler);
-    
+
     /* TI1 Configuration */
     TI1_Config(icpolarity, icselection, TIM1_ICFilter);
-    
+
     /* Set the Input Capture Prescaler value */
     TIM1_SetIC1Prescaler(TIM1_ICPrescaler);
-  }
-}
-
-/**
-  * @brief  Enables or disables the TIM1 peripheral.
-  * @param  NewState new state of the TIM1 peripheral.
-  *         This parameter can be ENABLE or DISABLE.
-  * @retval None
-  */
-void TIM1_Cmd(FunctionalState NewState)
-{
-  /* Check the parameters */
-  assert_param(IS_FUNCTIONALSTATE_OK(NewState));
-  
-  /* set or Reset the CEN Bit */
-  if (NewState != DISABLE)
-  {
-    TIM1->CR1 |= TIM1_CR1_CEN;
-  }
-  else
-  {
-    TIM1->CR1 &= (uint8_t)(~TIM1_CR1_CEN);
   }
 }
 
@@ -584,9 +530,9 @@ void TIM1_CtrlPWMOutputs(FunctionalState NewState)
 {
   /* Check the parameters */
   assert_param(IS_FUNCTIONALSTATE_OK(NewState));
-  
+
   /* Set or Reset the MOE Bit */
-  
+
   if (NewState != DISABLE)
   {
     TIM1->BKR |= TIM1_BKR_MOE;
@@ -594,41 +540,6 @@ void TIM1_CtrlPWMOutputs(FunctionalState NewState)
   else
   {
     TIM1->BKR &= (uint8_t)(~TIM1_BKR_MOE);
-  }
-}
-
-/**
-  * @brief  Enables or disables the specified TIM1 interrupts.
-  * @param  NewState new state of the TIM1 peripheral.
-  *         This parameter can be: ENABLE or DISABLE.
-  * @param  TIM1_IT specifies the TIM1 interrupts sources to be enabled or disabled.
-  *         This parameter can be any combination of the following values:
-  *           - TIM1_IT_UPDATE: TIM1 update Interrupt source
-  *           - TIM1_IT_CC1: TIM1 Capture Compare 1 Interrupt source
-  *           - TIM1_IT_CC2: TIM1 Capture Compare 2 Interrupt source
-  *           - TIM1_IT_CC3: TIM1 Capture Compare 3 Interrupt source
-  *           - TIM1_IT_CC4: TIM1 Capture Compare 4 Interrupt source
-  *           - TIM1_IT_CCUpdate: TIM1 Capture Compare Update Interrupt source
-  *           - TIM1_IT_TRIGGER: TIM1 Trigger Interrupt source
-  *           - TIM1_IT_BREAK: TIM1 Break Interrupt source
-  * @param  NewState new state of the TIM1 peripheral.
-  * @retval None
-  */
-void TIM1_ITConfig(TIM1_IT_TypeDef  TIM1_IT, FunctionalState NewState)
-{
-  /* Check the parameters */
-  assert_param(IS_TIM1_IT_OK(TIM1_IT));
-  assert_param(IS_FUNCTIONALSTATE_OK(NewState));
-  
-  if (NewState != DISABLE)
-  {
-    /* Enable the Interrupt sources */
-    TIM1->IER |= (uint8_t)TIM1_IT;
-  }
-  else
-  {
-    /* Disable the Interrupt sources */
-    TIM1->IER &= (uint8_t)(~(uint8_t)TIM1_IT);
   }
 }
 
@@ -666,10 +577,10 @@ void TIM1_ETRClockMode1Config(TIM1_ExtTRGPSC_TypeDef TIM1_ExtTRGPrescaler,
   /* Check the parameters */
   assert_param(IS_TIM1_EXT_PRESCALER_OK(TIM1_ExtTRGPrescaler));
   assert_param(IS_TIM1_EXT_POLARITY_OK(TIM1_ExtTRGPolarity));
-  
+
   /* Configure the ETR Clock source */
   TIM1_ETRConfig(TIM1_ExtTRGPrescaler, TIM1_ExtTRGPolarity, ExtTRGFilter);
-  
+
   /* Select the External clock mode1 & Select the Trigger selection : ETRF */
   TIM1->SMCR = (uint8_t)((uint8_t)(TIM1->SMCR & (uint8_t)(~(uint8_t)(TIM1_SMCR_SMS | TIM1_SMCR_TS )))
                          | (uint8_t)((uint8_t)TIM1_SLAVEMODE_EXTERNAL1 | TIM1_TS_ETRF ));
@@ -698,10 +609,10 @@ void TIM1_ETRClockMode2Config(TIM1_ExtTRGPSC_TypeDef TIM1_ExtTRGPrescaler,
   /* Check the parameters */
   assert_param(IS_TIM1_EXT_PRESCALER_OK(TIM1_ExtTRGPrescaler));
   assert_param(IS_TIM1_EXT_POLARITY_OK(TIM1_ExtTRGPolarity));
-  
+
   /* Configure the ETR Clock source */
   TIM1_ETRConfig(TIM1_ExtTRGPrescaler, TIM1_ExtTRGPolarity, ExtTRGFilter);
-  
+
   /* Enable the External clock mode2 */
   TIM1->ETR |= TIM1_ETR_ECE;
 }
@@ -756,7 +667,7 @@ void TIM1_TIxExternalClockConfig(TIM1_TIxExternalCLK1Source_TypeDef TIM1_TIxExte
   assert_param(IS_TIM1_TIXCLK_SOURCE_OK(TIM1_TIxExternalCLKSource));
   assert_param(IS_TIM1_IC_POLARITY_OK(TIM1_ICPolarity));
   assert_param(IS_TIM1_IC_FILTER_OK(ICFilter));
-  
+
   /* Configure the TIM1 Input Clock Source */
   if (TIM1_TIxExternalCLKSource == TIM1_TIXEXTERNALCLK1SOURCE_TI2)
   {
@@ -766,10 +677,10 @@ void TIM1_TIxExternalClockConfig(TIM1_TIxExternalCLK1Source_TypeDef TIM1_TIxExte
   {
     TI1_Config((uint8_t)TIM1_ICPolarity, (uint8_t)TIM1_ICSELECTION_DIRECTTI, (uint8_t)ICFilter);
   }
-  
+
   /* Select the Trigger source */
   TIM1_SelectInputTrigger((TIM1_TS_TypeDef)TIM1_TIxExternalCLKSource);
-  
+
   /* Select the External clock mode1 */
   TIM1->SMCR |= (uint8_t)(TIM1_SLAVEMODE_EXTERNAL1);
 }
@@ -788,7 +699,7 @@ void TIM1_SelectInputTrigger(TIM1_TS_TypeDef TIM1_InputTriggerSource)
 {
   /* Check the parameters */
   assert_param(IS_TIM1_TRIGGER_SELECTION_OK(TIM1_InputTriggerSource));
-  
+
   /* Select the Tgigger Source */
   TIM1->SMCR = (uint8_t)((uint8_t)(TIM1->SMCR & (uint8_t)(~TIM1_SMCR_TS)) | (uint8_t)TIM1_InputTriggerSource);
 }
@@ -804,7 +715,7 @@ void TIM1_UpdateDisableConfig(FunctionalState NewState)
 {
   /* Check the parameters */
   assert_param(IS_FUNCTIONALSTATE_OK(NewState));
-  
+
   /* Set or Reset the UDIS Bit */
   if (NewState != DISABLE)
   {
@@ -828,7 +739,7 @@ void TIM1_UpdateRequestConfig(TIM1_UpdateSource_TypeDef TIM1_UpdateSource)
 {
   /* Check the parameters */
   assert_param(IS_TIM1_UPDATE_SOURCE_OK(TIM1_UpdateSource));
-  
+
   /* Set or Reset the URS Bit */
   if (TIM1_UpdateSource != TIM1_UPDATESOURCE_GLOBAL)
   {
@@ -850,7 +761,7 @@ void TIM1_SelectHallSensor(FunctionalState NewState)
 {
   /* Check the parameters */
   assert_param(IS_FUNCTIONALSTATE_OK(NewState));
-  
+
   /* Set or Reset the TI1S Bit */
   if (NewState != DISABLE)
   {
@@ -874,7 +785,7 @@ void TIM1_SelectOnePulseMode(TIM1_OPMode_TypeDef TIM1_OPMode)
 {
   /* Check the parameters */
   assert_param(IS_TIM1_OPM_MODE_OK(TIM1_OPMode));
-  
+
   /* Set or Reset the OPM Bit */
   if (TIM1_OPMode != TIM1_OPMODE_REPETITIVE)
   {
@@ -884,7 +795,7 @@ void TIM1_SelectOnePulseMode(TIM1_OPMode_TypeDef TIM1_OPMode)
   {
     TIM1->CR1 &= (uint8_t)(~TIM1_CR1_OPM);
   }
-  
+
 }
 
 /**
@@ -904,9 +815,9 @@ void TIM1_SelectOutputTrigger(TIM1_TRGOSource_TypeDef TIM1_TRGOSource)
 {
   /* Check the parameters */
   assert_param(IS_TIM1_TRGO_SOURCE_OK(TIM1_TRGOSource));
-  
+
   /* Reset the MMS Bits & Select the TRGO source */
-  TIM1->CR2 = (uint8_t)((uint8_t)(TIM1->CR2 & (uint8_t)(~TIM1_CR2_MMS)) | 
+  TIM1->CR2 = (uint8_t)((uint8_t)(TIM1->CR2 & (uint8_t)(~TIM1_CR2_MMS)) |
                         (uint8_t) TIM1_TRGOSource);
 }
 
@@ -924,7 +835,7 @@ void TIM1_SelectSlaveMode(TIM1_SlaveMode_TypeDef TIM1_SlaveMode)
 {
   /* Check the parameters */
   assert_param(IS_TIM1_SLAVE_MODE_OK(TIM1_SlaveMode));
-  
+
   /* Reset the SMS Bits */ /* Select the Slave Mode */
   TIM1->SMCR = (uint8_t)((uint8_t)(TIM1->SMCR & (uint8_t)(~TIM1_SMCR_SMS)) |
                          (uint8_t)TIM1_SlaveMode);
@@ -940,7 +851,7 @@ void TIM1_SelectMasterSlaveMode(FunctionalState NewState)
 {
   /* Check the parameters */
   assert_param(IS_FUNCTIONALSTATE_OK(NewState));
-  
+
   /* Set or Reset the MSM Bit */
   if (NewState != DISABLE)
   {
@@ -980,7 +891,7 @@ void TIM1_EncoderInterfaceConfig(TIM1_EncoderMode_TypeDef TIM1_EncoderMode,
   assert_param(IS_TIM1_ENCODER_MODE_OK(TIM1_EncoderMode));
   assert_param(IS_TIM1_IC_POLARITY_OK(TIM1_IC1Polarity));
   assert_param(IS_TIM1_IC_POLARITY_OK(TIM1_IC2Polarity));
-  
+
   /* Set the TI1 and the TI2 Polarities */
   if (TIM1_IC1Polarity != TIM1_ICPOLARITY_RISING)
   {
@@ -990,7 +901,7 @@ void TIM1_EncoderInterfaceConfig(TIM1_EncoderMode_TypeDef TIM1_EncoderMode,
   {
     TIM1->CCER1 &= (uint8_t)(~TIM1_CCER1_CC1P);
   }
-  
+
   if (TIM1_IC2Polarity != TIM1_ICPOLARITY_RISING)
   {
     TIM1->CCER1 |= TIM1_CCER1_CC2P;
@@ -1002,9 +913,9 @@ void TIM1_EncoderInterfaceConfig(TIM1_EncoderMode_TypeDef TIM1_EncoderMode,
   /* Set the encoder Mode */
   TIM1->SMCR = (uint8_t)((uint8_t)(TIM1->SMCR & (uint8_t)(TIM1_SMCR_MSM | TIM1_SMCR_TS))
                          | (uint8_t) TIM1_EncoderMode);
-  
+
   /* Select the Capture Compare 1 and the Capture Compare 2 as input */
-  TIM1->CCMR1 = (uint8_t)((uint8_t)(TIM1->CCMR1 & (uint8_t)(~TIM1_CCMR_CCxS)) 
+  TIM1->CCMR1 = (uint8_t)((uint8_t)(TIM1->CCMR1 & (uint8_t)(~TIM1_CCMR_CCxS))
                           | (uint8_t) CCMR_TIxDirect_Set);
   TIM1->CCMR2 = (uint8_t)((uint8_t)(TIM1->CCMR2 & (uint8_t)(~TIM1_CCMR_CCxS))
                           | (uint8_t) CCMR_TIxDirect_Set);
@@ -1025,11 +936,11 @@ void TIM1_PrescalerConfig(uint16_t Prescaler,
 {
   /* Check the parameters */
   assert_param(IS_TIM1_PRESCALER_RELOAD_OK(TIM1_PSCReloadMode));
-  
+
   /* Set the Prescaler value */
   TIM1->PSCRH = (uint8_t)(Prescaler >> 8);
   TIM1->PSCRL = (uint8_t)(Prescaler);
-  
+
   /* Set or reset the UG Bit */
   TIM1->EGR = (uint8_t)TIM1_PSCReloadMode;
 }
@@ -1049,8 +960,8 @@ void TIM1_CounterModeConfig(TIM1_CounterMode_TypeDef TIM1_CounterMode)
 {
   /* Check the parameters */
   assert_param(IS_TIM1_COUNTER_MODE_OK(TIM1_CounterMode));
-  
-  
+
+
   /* Reset the CMS and DIR Bits & Set the Counter Mode */
   TIM1->CR1 = (uint8_t)((uint8_t)(TIM1->CR1 & (uint8_t)((uint8_t)(~TIM1_CR1_CMS) & (uint8_t)(~TIM1_CR1_DIR)))
                         | (uint8_t)TIM1_CounterMode);
@@ -1068,7 +979,7 @@ void TIM1_ForcedOC1Config(TIM1_ForcedAction_TypeDef TIM1_ForcedAction)
 {
   /* Check the parameters */
   assert_param(IS_TIM1_FORCED_ACTION_OK(TIM1_ForcedAction));
-  
+
   /* Reset the OCM Bits & Configure the Forced output Mode */
   TIM1->CCMR1 =  (uint8_t)((uint8_t)(TIM1->CCMR1 & (uint8_t)(~TIM1_CCMR_OCM))|
                            (uint8_t)TIM1_ForcedAction);
@@ -1086,7 +997,7 @@ void TIM1_ForcedOC2Config(TIM1_ForcedAction_TypeDef TIM1_ForcedAction)
 {
   /* Check the parameters */
   assert_param(IS_TIM1_FORCED_ACTION_OK(TIM1_ForcedAction));
-  
+
   /* Reset the OCM Bits & Configure the Forced output Mode */
   TIM1->CCMR2  =  (uint8_t)((uint8_t)(TIM1->CCMR2 & (uint8_t)(~TIM1_CCMR_OCM))
                             | (uint8_t)TIM1_ForcedAction);
@@ -1105,9 +1016,9 @@ void TIM1_ForcedOC3Config(TIM1_ForcedAction_TypeDef TIM1_ForcedAction)
 {
   /* Check the parameters */
   assert_param(IS_TIM1_FORCED_ACTION_OK(TIM1_ForcedAction));
-  
+
   /* Reset the OCM Bits */ /* Configure The Forced output Mode */
-  TIM1->CCMR3  =  (uint8_t)((uint8_t)(TIM1->CCMR3 & (uint8_t)(~TIM1_CCMR_OCM))  
+  TIM1->CCMR3  =  (uint8_t)((uint8_t)(TIM1->CCMR3 & (uint8_t)(~TIM1_CCMR_OCM))
                             | (uint8_t)TIM1_ForcedAction);
 }
 
@@ -1124,9 +1035,9 @@ void TIM1_ForcedOC4Config(TIM1_ForcedAction_TypeDef TIM1_ForcedAction)
 {
   /* Check the parameters */
   assert_param(IS_TIM1_FORCED_ACTION_OK(TIM1_ForcedAction));
-  
+
   /* Reset the OCM Bits & Configure the Forced output Mode */
-  TIM1->CCMR4  =  (uint8_t)((uint8_t)(TIM1->CCMR4 & (uint8_t)(~TIM1_CCMR_OCM)) 
+  TIM1->CCMR4  =  (uint8_t)((uint8_t)(TIM1->CCMR4 & (uint8_t)(~TIM1_CCMR_OCM))
                             | (uint8_t)TIM1_ForcedAction);
 }
 
@@ -1140,7 +1051,7 @@ void TIM1_ARRPreloadConfig(FunctionalState NewState)
 {
   /* Check the parameters */
   assert_param(IS_FUNCTIONALSTATE_OK(NewState));
-  
+
   /* Set or Reset the ARPE Bit */
   if (NewState != DISABLE)
   {
@@ -1162,7 +1073,7 @@ void TIM1_SelectCOM(FunctionalState NewState)
 {
   /* Check the parameters */
   assert_param(IS_FUNCTIONALSTATE_OK(NewState));
-  
+
   /* Set or Reset the COMS Bit */
   if (NewState != DISABLE)
   {
@@ -1184,7 +1095,7 @@ void TIM1_CCPreloadControl(FunctionalState NewState)
 {
   /* Check the parameters */
   assert_param(IS_FUNCTIONALSTATE_OK(NewState));
-  
+
   /* Set or Reset the CCPC Bit */
   if (NewState != DISABLE)
   {
@@ -1206,7 +1117,7 @@ void TIM1_OC1PreloadConfig(FunctionalState NewState)
 {
   /* Check the parameters */
   assert_param(IS_FUNCTIONALSTATE_OK(NewState));
-  
+
   /* Set or Reset the OC1PE Bit */
   if (NewState != DISABLE)
   {
@@ -1228,7 +1139,7 @@ void TIM1_OC2PreloadConfig(FunctionalState NewState)
 {
   /* Check the parameters */
   assert_param(IS_FUNCTIONALSTATE_OK(NewState));
-  
+
   /* Set or Reset the OC2PE Bit */
   if (NewState != DISABLE)
   {
@@ -1250,7 +1161,7 @@ void TIM1_OC3PreloadConfig(FunctionalState NewState)
 {
   /* Check the parameters */
   assert_param(IS_FUNCTIONALSTATE_OK(NewState));
-  
+
   /* Set or Reset the OC3PE Bit */
   if (NewState != DISABLE)
   {
@@ -1272,7 +1183,7 @@ void TIM1_OC4PreloadConfig(FunctionalState NewState)
 {
   /* Check the parameters */
   assert_param(IS_FUNCTIONALSTATE_OK(NewState));
-  
+
   /* Set or Reset the OC4PE Bit */
   if (NewState != DISABLE)
   {
@@ -1294,7 +1205,7 @@ void TIM1_OC1FastConfig(FunctionalState NewState)
 {
   /* Check the parameters */
   assert_param(IS_FUNCTIONALSTATE_OK(NewState));
-  
+
   /* Set or Reset the OC1FE Bit */
   if (NewState != DISABLE)
   {
@@ -1316,7 +1227,7 @@ void TIM1_OC2FastConfig(FunctionalState NewState)
 {
   /* Check the parameters */
   assert_param(IS_FUNCTIONALSTATE_OK(NewState));
-  
+
   /* Set or Reset the OC2FE Bit */
   if (NewState != DISABLE)
   {
@@ -1338,7 +1249,7 @@ void TIM1_OC3FastConfig(FunctionalState NewState)
 {
   /* Check the parameters */
   assert_param(IS_FUNCTIONALSTATE_OK(NewState));
-  
+
   /* Set or Reset the OC3FE Bit */
   if (NewState != DISABLE)
   {
@@ -1360,7 +1271,7 @@ void TIM1_OC4FastConfig(FunctionalState NewState)
 {
   /* Check the parameters */
   assert_param(IS_FUNCTIONALSTATE_OK(NewState));
-  
+
   /* Set or Reset the OC4FE Bit */
   if (NewState != DISABLE)
   {
@@ -1390,7 +1301,7 @@ void TIM1_GenerateEvent(TIM1_EventSource_TypeDef TIM1_EventSource)
 {
   /* Check the parameters */
   assert_param(IS_TIM1_EVENT_SOURCE_OK(TIM1_EventSource));
-  
+
   /* Set the event sources */
   TIM1->EGR = (uint8_t)TIM1_EventSource;
 }
@@ -1407,7 +1318,7 @@ void TIM1_OC1PolarityConfig(TIM1_OCPolarity_TypeDef TIM1_OCPolarity)
 {
   /* Check the parameters */
   assert_param(IS_TIM1_OC_POLARITY_OK(TIM1_OCPolarity));
-  
+
   /* Set or Reset the CC1P Bit */
   if (TIM1_OCPolarity != TIM1_OCPOLARITY_HIGH)
   {
@@ -1431,7 +1342,7 @@ void TIM1_OC1NPolarityConfig(TIM1_OCNPolarity_TypeDef TIM1_OCNPolarity)
 {
   /* Check the parameters */
   assert_param(IS_TIM1_OCN_POLARITY_OK(TIM1_OCNPolarity));
-  
+
   /* Set or Reset the CC3P Bit */
   if (TIM1_OCNPolarity != TIM1_OCNPOLARITY_HIGH)
   {
@@ -1455,7 +1366,7 @@ void TIM1_OC2PolarityConfig(TIM1_OCPolarity_TypeDef TIM1_OCPolarity)
 {
   /* Check the parameters */
   assert_param(IS_TIM1_OC_POLARITY_OK(TIM1_OCPolarity));
-  
+
   /* Set or Reset the CC2P Bit */
   if (TIM1_OCPolarity != TIM1_OCPOLARITY_HIGH)
   {
@@ -1479,7 +1390,7 @@ void TIM1_OC2NPolarityConfig(TIM1_OCNPolarity_TypeDef TIM1_OCNPolarity)
 {
   /* Check the parameters */
   assert_param(IS_TIM1_OCN_POLARITY_OK(TIM1_OCNPolarity));
-  
+
   /* Set or Reset the CC3P Bit */
   if (TIM1_OCNPolarity != TIM1_OCNPOLARITY_HIGH)
   {
@@ -1503,7 +1414,7 @@ void TIM1_OC3PolarityConfig(TIM1_OCPolarity_TypeDef TIM1_OCPolarity)
 {
   /* Check the parameters */
   assert_param(IS_TIM1_OC_POLARITY_OK(TIM1_OCPolarity));
-  
+
   /* Set or Reset the CC3P Bit */
   if (TIM1_OCPolarity != TIM1_OCPOLARITY_HIGH)
   {
@@ -1528,7 +1439,7 @@ void TIM1_OC3NPolarityConfig(TIM1_OCNPolarity_TypeDef TIM1_OCNPolarity)
 {
   /* Check the parameters */
   assert_param(IS_TIM1_OCN_POLARITY_OK(TIM1_OCNPolarity));
-  
+
   /* Set or Reset the CC3P Bit */
   if (TIM1_OCNPolarity != TIM1_OCNPOLARITY_HIGH)
   {
@@ -1552,7 +1463,7 @@ void TIM1_OC4PolarityConfig(TIM1_OCPolarity_TypeDef TIM1_OCPolarity)
 {
   /* Check the parameters */
   assert_param(IS_TIM1_OC_POLARITY_OK(TIM1_OCPolarity));
-  
+
   /* Set or Reset the CC4P Bit */
   if (TIM1_OCPolarity != TIM1_OCPOLARITY_HIGH)
   {
@@ -1581,7 +1492,7 @@ void TIM1_CCxCmd(TIM1_Channel_TypeDef TIM1_Channel, FunctionalState NewState)
   /* Check the parameters */
   assert_param(IS_TIM1_CHANNEL_OK(TIM1_Channel));
   assert_param(IS_FUNCTIONALSTATE_OK(NewState));
-  
+
   if (TIM1_Channel == TIM1_CHANNEL_1)
   {
     /* Set or Reset the CC1E Bit */
@@ -1593,7 +1504,7 @@ void TIM1_CCxCmd(TIM1_Channel_TypeDef TIM1_Channel, FunctionalState NewState)
     {
       TIM1->CCER1 &= (uint8_t)(~TIM1_CCER1_CC1E);
     }
-    
+
   }
   else if (TIM1_Channel == TIM1_CHANNEL_2)
   {
@@ -1649,7 +1560,7 @@ void TIM1_CCxNCmd(TIM1_Channel_TypeDef TIM1_Channel, FunctionalState NewState)
   /* Check the parameters */
   assert_param(IS_TIM1_COMPLEMENTARY_CHANNEL_OK(TIM1_Channel));
   assert_param(IS_FUNCTIONALSTATE_OK(NewState));
-  
+
   if (TIM1_Channel == TIM1_CHANNEL_1)
   {
     /* Set or Reset the CC1NE Bit */
@@ -1714,21 +1625,21 @@ void TIM1_SelectOCxM(TIM1_Channel_TypeDef TIM1_Channel, TIM1_OCMode_TypeDef TIM1
   /* Check the parameters */
   assert_param(IS_TIM1_CHANNEL_OK(TIM1_Channel));
   assert_param(IS_TIM1_OCM_OK(TIM1_OCMode));
-  
+
   if (TIM1_Channel == TIM1_CHANNEL_1)
   {
     /* Disable the Channel 1: Reset the CCE Bit */
     TIM1->CCER1 &= (uint8_t)(~TIM1_CCER1_CC1E);
-    
+
     /* Reset the Output Compare Bits & Set the Output Compare Mode */
-    TIM1->CCMR1 = (uint8_t)((uint8_t)(TIM1->CCMR1 & (uint8_t)(~TIM1_CCMR_OCM)) 
+    TIM1->CCMR1 = (uint8_t)((uint8_t)(TIM1->CCMR1 & (uint8_t)(~TIM1_CCMR_OCM))
                             | (uint8_t)TIM1_OCMode);
   }
   else if (TIM1_Channel == TIM1_CHANNEL_2)
   {
     /* Disable the Channel 2: Reset the CCE Bit */
     TIM1->CCER1 &= (uint8_t)(~TIM1_CCER1_CC2E);
-    
+
     /* Reset the Output Compare Bits & Set the Output Compare Mode */
     TIM1->CCMR2 = (uint8_t)((uint8_t)(TIM1->CCMR2 & (uint8_t)(~TIM1_CCMR_OCM))
                             | (uint8_t)TIM1_OCMode);
@@ -1737,18 +1648,18 @@ void TIM1_SelectOCxM(TIM1_Channel_TypeDef TIM1_Channel, TIM1_OCMode_TypeDef TIM1
   {
     /* Disable the Channel 3: Reset the CCE Bit */
     TIM1->CCER2 &= (uint8_t)(~TIM1_CCER2_CC3E);
-    
+
     /* Reset the Output Compare Bits & Set the Output Compare Mode */
-    TIM1->CCMR3 = (uint8_t)((uint8_t)(TIM1->CCMR3 & (uint8_t)(~TIM1_CCMR_OCM)) 
+    TIM1->CCMR3 = (uint8_t)((uint8_t)(TIM1->CCMR3 & (uint8_t)(~TIM1_CCMR_OCM))
                             | (uint8_t)TIM1_OCMode);
   }
   else
   {
     /* Disable the Channel 4: Reset the CCE Bit */
     TIM1->CCER2 &= (uint8_t)(~TIM1_CCER2_CC4E);
-    
+
     /* Reset the Output Compare Bits & Set the Output Compare Mode */
-    TIM1->CCMR4 = (uint8_t)((uint8_t)(TIM1->CCMR4 & (uint8_t)(~TIM1_CCMR_OCM)) 
+    TIM1->CCMR4 = (uint8_t)((uint8_t)(TIM1->CCMR4 & (uint8_t)(~TIM1_CCMR_OCM))
                             | (uint8_t)TIM1_OCMode);
   }
 }
@@ -1845,9 +1756,9 @@ void TIM1_SetIC1Prescaler(TIM1_ICPSC_TypeDef TIM1_IC1Prescaler)
 {
   /* Check the parameters */
   assert_param(IS_TIM1_IC_PRESCALER_OK(TIM1_IC1Prescaler));
-  
+
   /* Reset the IC1PSC Bits */ /* Set the IC1PSC value */
-  TIM1->CCMR1 = (uint8_t)((uint8_t)(TIM1->CCMR1 & (uint8_t)(~TIM1_CCMR_ICxPSC)) 
+  TIM1->CCMR1 = (uint8_t)((uint8_t)(TIM1->CCMR1 & (uint8_t)(~TIM1_CCMR_ICxPSC))
                           | (uint8_t)TIM1_IC1Prescaler);
 }
 
@@ -1863,10 +1774,10 @@ void TIM1_SetIC1Prescaler(TIM1_ICPSC_TypeDef TIM1_IC1Prescaler)
   */
 void TIM1_SetIC2Prescaler(TIM1_ICPSC_TypeDef TIM1_IC2Prescaler)
 {
-  
+
   /* Check the parameters */
   assert_param(IS_TIM1_IC_PRESCALER_OK(TIM1_IC2Prescaler));
-  
+
   /* Reset the IC1PSC Bits */ /* Set the IC1PSC value */
   TIM1->CCMR2 = (uint8_t)((uint8_t)(TIM1->CCMR2 & (uint8_t)(~TIM1_CCMR_ICxPSC))
                           | (uint8_t)TIM1_IC2Prescaler);
@@ -1884,12 +1795,12 @@ void TIM1_SetIC2Prescaler(TIM1_ICPSC_TypeDef TIM1_IC2Prescaler)
   */
 void TIM1_SetIC3Prescaler(TIM1_ICPSC_TypeDef TIM1_IC3Prescaler)
 {
-  
+
   /* Check the parameters */
   assert_param(IS_TIM1_IC_PRESCALER_OK(TIM1_IC3Prescaler));
-  
+
   /* Reset the IC1PSC Bits & Set the IC1PSC value */
-  TIM1->CCMR3 = (uint8_t)((uint8_t)(TIM1->CCMR3 & (uint8_t)(~TIM1_CCMR_ICxPSC)) | 
+  TIM1->CCMR3 = (uint8_t)((uint8_t)(TIM1->CCMR3 & (uint8_t)(~TIM1_CCMR_ICxPSC)) |
                           (uint8_t)TIM1_IC3Prescaler);
 }
 
@@ -1905,10 +1816,10 @@ void TIM1_SetIC3Prescaler(TIM1_ICPSC_TypeDef TIM1_IC3Prescaler)
   */
 void TIM1_SetIC4Prescaler(TIM1_ICPSC_TypeDef TIM1_IC4Prescaler)
 {
-  
+
   /* Check the parameters */
   assert_param(IS_TIM1_IC_PRESCALER_OK(TIM1_IC4Prescaler));
-  
+
   /* Reset the IC1PSC Bits &  Set the IC1PSC value */
   TIM1->CCMR4 = (uint8_t)((uint8_t)(TIM1->CCMR4 & (uint8_t)(~TIM1_CCMR_ICxPSC)) |
                           (uint8_t)TIM1_IC4Prescaler);
@@ -1922,13 +1833,13 @@ void TIM1_SetIC4Prescaler(TIM1_ICPSC_TypeDef TIM1_IC4Prescaler)
 uint16_t TIM1_GetCapture1(void)
 {
   /* Get the Capture 1 Register value */
-  
+
   uint16_t tmpccr1 = 0;
   uint8_t tmpccr1l=0, tmpccr1h=0;
-  
+
   tmpccr1h = TIM1->CCR1H;
   tmpccr1l = TIM1->CCR1L;
-  
+
   tmpccr1 = (uint16_t)(tmpccr1l);
   tmpccr1 |= (uint16_t)((uint16_t)tmpccr1h << 8);
   /* Get the Capture 1 Register value */
@@ -1943,13 +1854,13 @@ uint16_t TIM1_GetCapture1(void)
 uint16_t TIM1_GetCapture2(void)
 {
   /* Get the Capture 2 Register value */
-  
+
   uint16_t tmpccr2 = 0;
   uint8_t tmpccr2l=0, tmpccr2h=0;
-  
+
   tmpccr2h = TIM1->CCR2H;
   tmpccr2l = TIM1->CCR2L;
-  
+
   tmpccr2 = (uint16_t)(tmpccr2l);
   tmpccr2 |= (uint16_t)((uint16_t)tmpccr2h << 8);
   /* Get the Capture 2 Register value */
@@ -1966,10 +1877,10 @@ uint16_t TIM1_GetCapture3(void)
   /* Get the Capture 3 Register value */
   uint16_t tmpccr3 = 0;
   uint8_t tmpccr3l=0, tmpccr3h=0;
-  
+
   tmpccr3h = TIM1->CCR3H;
   tmpccr3l = TIM1->CCR3L;
-  
+
   tmpccr3 = (uint16_t)(tmpccr3l);
   tmpccr3 |= (uint16_t)((uint16_t)tmpccr3h << 8);
   /* Get the Capture 3 Register value */
@@ -1986,10 +1897,10 @@ uint16_t TIM1_GetCapture4(void)
   /* Get the Capture 4 Register value */
   uint16_t tmpccr4 = 0;
   uint8_t tmpccr4l=0, tmpccr4h=0;
-  
+
   tmpccr4h = TIM1->CCR4H;
   tmpccr4l = TIM1->CCR4L;
-  
+
   tmpccr4 = (uint16_t)(tmpccr4l);
   tmpccr4 |= (uint16_t)((uint16_t)tmpccr4h << 8);
   /* Get the Capture 4 Register value */
@@ -2004,9 +1915,9 @@ uint16_t TIM1_GetCapture4(void)
 uint16_t TIM1_GetCounter(void)
 {
   uint16_t tmpcntr = 0;
-  
+
   tmpcntr = ((uint16_t)TIM1->CNTRH << 8);
-  
+
   /* Get the Counter Register value */
   return (uint16_t)(tmpcntr | (uint16_t)(TIM1->CNTRL));
 }
@@ -2019,9 +1930,9 @@ uint16_t TIM1_GetCounter(void)
 uint16_t TIM1_GetPrescaler(void)
 {
   uint16_t temp = 0;
-  
+
   temp = ((uint16_t)TIM1->PSCRH << 8);
-  
+
   /* Get the Prescaler Register value */
   return (uint16_t)( temp | (uint16_t)(TIM1->PSCRL));
 }
@@ -2048,13 +1959,13 @@ FlagStatus TIM1_GetFlagStatus(TIM1_FLAG_TypeDef TIM1_FLAG)
 {
   FlagStatus bitstatus = RESET;
   uint8_t tim1_flag_l = 0, tim1_flag_h = 0;
-  
+
   /* Check the parameters */
   assert_param(IS_TIM1_GET_FLAG_OK(TIM1_FLAG));
-  
+
   tim1_flag_l = (uint8_t)(TIM1->SR1 & (uint8_t)TIM1_FLAG);
   tim1_flag_h = (uint8_t)((uint16_t)TIM1_FLAG >> 8);
-  
+
   if ((tim1_flag_l | (uint8_t)(TIM1->SR2 & tim1_flag_h)) != 0)
   {
     bitstatus = SET;
@@ -2088,10 +1999,10 @@ void TIM1_ClearFlag(TIM1_FLAG_TypeDef TIM1_FLAG)
 {
   /* Check the parameters */
   assert_param(IS_TIM1_CLEAR_FLAG_OK(TIM1_FLAG));
-  
+
   /* Clear the flags (rc_w0) clear this bit by writing 0. Writing ‘1’ has no effect*/
   TIM1->SR1 = (uint8_t)(~(uint8_t)(TIM1_FLAG));
-  TIM1->SR2 = (uint8_t)((uint8_t)(~((uint8_t)((uint16_t)TIM1_FLAG >> 8))) & 
+  TIM1->SR2 = (uint8_t)((uint8_t)(~((uint8_t)((uint16_t)TIM1_FLAG >> 8))) &
                         (uint8_t)0x1E);
 }
 
@@ -2113,14 +2024,14 @@ ITStatus TIM1_GetITStatus(TIM1_IT_TypeDef TIM1_IT)
 {
   ITStatus bitstatus = RESET;
   uint8_t TIM1_itStatus = 0, TIM1_itEnable = 0;
-  
+
   /* Check the parameters */
   assert_param(IS_TIM1_GET_IT_OK(TIM1_IT));
-  
+
   TIM1_itStatus = (uint8_t)(TIM1->SR1 & (uint8_t)TIM1_IT);
-  
+
   TIM1_itEnable = (uint8_t)(TIM1->IER & (uint8_t)TIM1_IT);
-  
+
   if ((TIM1_itStatus != (uint8_t)RESET ) && (TIM1_itEnable != (uint8_t)RESET ))
   {
     bitstatus = SET;
@@ -2130,29 +2041,6 @@ ITStatus TIM1_GetITStatus(TIM1_IT_TypeDef TIM1_IT)
     bitstatus = RESET;
   }
   return (ITStatus)(bitstatus);
-}
-
-/**
-  * @brief  Clears the TIM1's interrupt pending bits.
-  * @param  TIM1_IT specifies the pending bit to clear.
-  *         This parameter can be one of the following values:
-  *                       - TIM1_IT_UPDATE: TIM1 update Interrupt source
-  *                       - TIM1_IT_CC1: TIM1 Capture Compare 1 Interrupt source
-  *                       - TIM1_IT_CC2: TIM1 Capture Compare 2 Interrupt source
-  *                       - TIM1_IT_CC3: TIM1 Capture Compare 3 Interrupt source
-  *                       - TIM1_IT_CC4: TIM1 Capture Compare 4 Interrupt source
-  *                       - TIM1_IT_COM: TIM1 Commutation Interrupt source
-  *                       - TIM1_IT_TRIGGER: TIM1 Trigger Interrupt source
-  *                       - TIM1_IT_BREAK: TIM1 Break Interrupt source
-  * @retval None.
-  */
-void TIM1_ClearITPendingBit(TIM1_IT_TypeDef TIM1_IT)
-{
-  /* Check the parameters */
-  assert_param(IS_TIM1_IT_OK(TIM1_IT));
-  
-  /* Clear the IT pending Bit */
-  TIM1->SR1 = (uint8_t)(~(uint8_t)TIM1_IT);
 }
 
 /**
@@ -2177,11 +2065,11 @@ static void TI1_Config(uint8_t TIM1_ICPolarity,
 {
   /* Disable the Channel 1: Reset the CCE Bit */
   TIM1->CCER1 &= (uint8_t)(~TIM1_CCER1_CC1E);
-  
+
   /* Select the Input and set the filter */
-  TIM1->CCMR1 = (uint8_t)((uint8_t)(TIM1->CCMR1 & (uint8_t)(~(uint8_t)( TIM1_CCMR_CCxS | TIM1_CCMR_ICxF ))) | 
+  TIM1->CCMR1 = (uint8_t)((uint8_t)(TIM1->CCMR1 & (uint8_t)(~(uint8_t)( TIM1_CCMR_CCxS | TIM1_CCMR_ICxF ))) |
                           (uint8_t)(( (TIM1_ICSelection)) | ((uint8_t)( TIM1_ICFilter << 4))));
-  
+
   /* Select the Polarity */
   if (TIM1_ICPolarity != TIM1_ICPOLARITY_RISING)
   {
@@ -2191,7 +2079,7 @@ static void TI1_Config(uint8_t TIM1_ICPolarity,
   {
     TIM1->CCER1 &= (uint8_t)(~TIM1_CCER1_CC1P);
   }
-  
+
   /* Set the CCE Bit */
   TIM1->CCER1 |=  TIM1_CCER1_CC1E;
 }
@@ -2218,9 +2106,9 @@ static void TI2_Config(uint8_t TIM1_ICPolarity,
 {
   /* Disable the Channel 2: Reset the CCE Bit */
   TIM1->CCER1 &=  (uint8_t)(~TIM1_CCER1_CC2E);
-  
+
   /* Select the Input and set the filter */
-  TIM1->CCMR2  = (uint8_t)((uint8_t)(TIM1->CCMR2 & (uint8_t)(~(uint8_t)( TIM1_CCMR_CCxS | TIM1_CCMR_ICxF ))) 
+  TIM1->CCMR2  = (uint8_t)((uint8_t)(TIM1->CCMR2 & (uint8_t)(~(uint8_t)( TIM1_CCMR_CCxS | TIM1_CCMR_ICxF )))
                            | (uint8_t)(( (TIM1_ICSelection)) | ((uint8_t)( TIM1_ICFilter << 4))));
   /* Select the Polarity */
   if (TIM1_ICPolarity != TIM1_ICPOLARITY_RISING)
@@ -2257,11 +2145,11 @@ static void TI3_Config(uint8_t TIM1_ICPolarity,
 {
   /* Disable the Channel 3: Reset the CCE Bit */
   TIM1->CCER2 &=  (uint8_t)(~TIM1_CCER2_CC3E);
-  
+
   /* Select the Input and set the filter */
-  TIM1->CCMR3 = (uint8_t)((uint8_t)(TIM1->CCMR3 & (uint8_t)(~(uint8_t)( TIM1_CCMR_CCxS | TIM1_CCMR_ICxF))) 
+  TIM1->CCMR3 = (uint8_t)((uint8_t)(TIM1->CCMR3 & (uint8_t)(~(uint8_t)( TIM1_CCMR_CCxS | TIM1_CCMR_ICxF)))
                           | (uint8_t)(( (TIM1_ICSelection)) | ((uint8_t)( TIM1_ICFilter << 4))));
-  
+
   /* Select the Polarity */
   if (TIM1_ICPolarity != TIM1_ICPOLARITY_RISING)
   {
@@ -2297,11 +2185,11 @@ static void TI4_Config(uint8_t TIM1_ICPolarity,
 {
   /* Disable the Channel 4: Reset the CCE Bit */
   TIM1->CCER2 &=  (uint8_t)(~TIM1_CCER2_CC4E);
-  
+
   /* Select the Input and set the filter */
   TIM1->CCMR4 = (uint8_t)((uint8_t)(TIM1->CCMR4 & (uint8_t)(~(uint8_t)( TIM1_CCMR_CCxS | TIM1_CCMR_ICxF )))
                           | (uint8_t)(( (TIM1_ICSelection)) | ((uint8_t)( TIM1_ICFilter << 4))));
-  
+
   /* Select the Polarity */
   if (TIM1_ICPolarity != TIM1_ICPOLARITY_RISING)
   {
@@ -2311,7 +2199,7 @@ static void TI4_Config(uint8_t TIM1_ICPolarity,
   {
     TIM1->CCER2 &= (uint8_t)(~TIM1_CCER2_CC4P);
   }
-  
+
   /* Set the CCE Bit */
   TIM1->CCER2 |=  TIM1_CCER2_CC4E;
 }
@@ -2320,10 +2208,10 @@ static void TI4_Config(uint8_t TIM1_ICPolarity,
 /**
   * @}
   */
-  
+
 /**
   * @}
   */
-  
+
 
 /************************ (C) COPYRIGHT STMicroelectronics *****END OF FILE****/
