+++ src/stm8s_uart1.c
@@ -16,8 +16,8 @@
   *
   *        http://www.st.com/software_license_agreement_liberty_v2
   *
-  * Unless required by applicable law or agreed to in writing, software 
-  * distributed under the License is distributed on an "AS IS" BASIS, 
+  * Unless required by applicable law or agreed to in writing, software
+  * distributed under the License is distributed on an "AS IS" BASIS,
   * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
   * See the License for the specific language governing permissions and
   * limitations under the License.
@@ -56,204 +56,18 @@ void UART1_DeInit(void)
   to the UART1_SR register followed by a Read to the UART1_DR register */
   (void)UART1->SR;
   (void)UART1->DR;
-  
+
   UART1->BRR2 = UART1_BRR2_RESET_VALUE;  /* Set UART1_BRR2 to reset value 0x00 */
   UART1->BRR1 = UART1_BRR1_RESET_VALUE;  /* Set UART1_BRR1 to reset value 0x00 */
-  
+
   UART1->CR1 = UART1_CR1_RESET_VALUE;  /* Set UART1_CR1 to reset value 0x00 */
   UART1->CR2 = UART1_CR2_RESET_VALUE;  /* Set UART1_CR2 to reset value 0x00 */
   UART1->CR3 = UART1_CR3_RESET_VALUE;  /* Set UART1_CR3 to reset value 0x00 */
   UART1->CR4 = UART1_CR4_RESET_VALUE;  /* Set UART1_CR4 to reset value 0x00 */
   UART1->CR5 = UART1_CR5_RESET_VALUE;  /* Set UART1_CR5 to reset value 0x00 */
-  
+
   UART1->GTR = UART1_GTR_RESET_VALUE;
   UART1->PSCR = UART1_PSCR_RESET_VALUE;
-}
-
-/**
-  * @brief  Initializes the UART1 according to the specified parameters.
-  * @note   Configure in Push Pull or Open Drain mode the Tx pin by setting the
-  *         correct I/O Port register according the product package and line
-  *         configuration
-  * @param  BaudRate: The baudrate.
-  * @param  WordLength : This parameter can be any of the 
-  *         @ref UART1_WordLength_TypeDef enumeration.
-  * @param  StopBits: This parameter can be any of the 
-  *         @ref UART1_StopBits_TypeDef enumeration.
-  * @param  Parity: This parameter can be any of the 
-  *         @ref UART1_Parity_TypeDef enumeration.
-  * @param  SyncMode: This parameter can be any of the 
-  *         @ref UART1_SyncMode_TypeDef values.
-  * @param  Mode: This parameter can be any of the @ref UART1_Mode_TypeDef values
-  * @retval None
-  */
-void UART1_Init(uint32_t BaudRate, UART1_WordLength_TypeDef WordLength, 
-                UART1_StopBits_TypeDef StopBits, UART1_Parity_TypeDef Parity, 
-                UART1_SyncMode_TypeDef SyncMode, UART1_Mode_TypeDef Mode)
-{
-  uint32_t BaudRate_Mantissa = 0, BaudRate_Mantissa100 = 0;
-  
-  /* Check the parameters */
-  assert_param(IS_UART1_BAUDRATE_OK(BaudRate));
-  assert_param(IS_UART1_WORDLENGTH_OK(WordLength));
-  assert_param(IS_UART1_STOPBITS_OK(StopBits));
-  assert_param(IS_UART1_PARITY_OK(Parity));
-  assert_param(IS_UART1_MODE_OK((uint8_t)Mode));
-  assert_param(IS_UART1_SYNCMODE_OK((uint8_t)SyncMode));
-  
-  /* Clear the word length bit */
-  UART1->CR1 &= (uint8_t)(~UART1_CR1_M);  
-  
-  /* Set the word length bit according to UART1_WordLength value */
-  UART1->CR1 |= (uint8_t)WordLength;
-  
-  /* Clear the STOP bits */
-  UART1->CR3 &= (uint8_t)(~UART1_CR3_STOP);  
-  /* Set the STOP bits number according to UART1_StopBits value  */
-  UART1->CR3 |= (uint8_t)StopBits;  
-  
-  /* Clear the Parity Control bit */
-  UART1->CR1 &= (uint8_t)(~(UART1_CR1_PCEN | UART1_CR1_PS  ));  
-  /* Set the Parity Control bit to UART1_Parity value */
-  UART1->CR1 |= (uint8_t)Parity;  
-  
-  /* Clear the LSB mantissa of UART1DIV  */
-  UART1->BRR1 &= (uint8_t)(~UART1_BRR1_DIVM);  
-  /* Clear the MSB mantissa of UART1DIV  */
-  UART1->BRR2 &= (uint8_t)(~UART1_BRR2_DIVM);  
-  /* Clear the Fraction bits of UART1DIV */
-  UART1->BRR2 &= (uint8_t)(~UART1_BRR2_DIVF);  
-  
-  /* Set the UART1 BaudRates in BRR1 and BRR2 registers according to UART1_BaudRate value */
-  BaudRate_Mantissa    = ((uint32_t)CLK_GetClockFreq() / (BaudRate << 4));
-  BaudRate_Mantissa100 = (((uint32_t)CLK_GetClockFreq() * 100) / (BaudRate << 4));
-  /* Set the fraction of UART1DIV  */
-  UART1->BRR2 |= (uint8_t)((uint8_t)(((BaudRate_Mantissa100 - (BaudRate_Mantissa * 100)) << 4) / 100) & (uint8_t)0x0F); 
-  /* Set the MSB mantissa of UART1DIV  */
-  UART1->BRR2 |= (uint8_t)((BaudRate_Mantissa >> 4) & (uint8_t)0xF0); 
-  /* Set the LSB mantissa of UART1DIV  */
-  UART1->BRR1 |= (uint8_t)BaudRate_Mantissa;           
-  
-  /* Disable the Transmitter and Receiver before setting the LBCL, CPOL and CPHA bits */
-  UART1->CR2 &= (uint8_t)~(UART1_CR2_TEN | UART1_CR2_REN); 
-  /* Clear the Clock Polarity, lock Phase, Last Bit Clock pulse */
-  UART1->CR3 &= (uint8_t)~(UART1_CR3_CPOL | UART1_CR3_CPHA | UART1_CR3_LBCL); 
-  /* Set the Clock Polarity, lock Phase, Last Bit Clock pulse */
-  UART1->CR3 |= (uint8_t)((uint8_t)SyncMode & (uint8_t)(UART1_CR3_CPOL | 
-                                                        UART1_CR3_CPHA | UART1_CR3_LBCL));  
-  
-  if ((uint8_t)(Mode & UART1_MODE_TX_ENABLE))
-  {
-    /* Set the Transmitter Enable bit */
-    UART1->CR2 |= (uint8_t)UART1_CR2_TEN;  
-  }
-  else
-  {
-    /* Clear the Transmitter Disable bit */
-    UART1->CR2 &= (uint8_t)(~UART1_CR2_TEN);  
-  }
-  if ((uint8_t)(Mode & UART1_MODE_RX_ENABLE))
-  {
-    /* Set the Receiver Enable bit */
-    UART1->CR2 |= (uint8_t)UART1_CR2_REN;  
-  }
-  else
-  {
-    /* Clear the Receiver Disable bit */
-    UART1->CR2 &= (uint8_t)(~UART1_CR2_REN);  
-  }
-  /* Set the Clock Enable bit, lock Polarity, lock Phase and Last Bit Clock 
-  pulse bits according to UART1_Mode value */
-  if ((uint8_t)(SyncMode & UART1_SYNCMODE_CLOCK_DISABLE))
-  {
-    /* Clear the Clock Enable bit */
-    UART1->CR3 &= (uint8_t)(~UART1_CR3_CKEN); 
-  }
-  else
-  {
-    UART1->CR3 |= (uint8_t)((uint8_t)SyncMode & UART1_CR3_CKEN);
-  }
-}
-
-/**
-  * @brief  Enable the UART1 peripheral.
-  * @param  NewState : The new state of the UART Communication.
-  *         This parameter can be any of the @ref FunctionalState enumeration.
-  * @retval None
-  */
-void UART1_Cmd(FunctionalState NewState)
-{
-  if (NewState != DISABLE)
-  {
-    /* UART1 Enable */
-    UART1->CR1 &= (uint8_t)(~UART1_CR1_UARTD); 
-  }
-  else
-  {
-    /* UART Disable */
-    UART1->CR1 |= UART1_CR1_UARTD;  
-  }
-}
-
-/**
-  * @brief  Enables or disables the specified USART interrupts.
-  * @param  UART1_IT specifies the USART interrupt sources to be enabled or disabled.
-  *         This parameter can be one of the following values:
-  *         - UART1_IT_TXE:  Transmit Data Register empty interrupt
-  *         - UART1_IT_TC:   Transmission complete interrupt
-  *         - UART1_IT_RXNE_OR: Receive Data register not empty and Overrun interrupt
-  *         - UART1_IT_IDLE: Idle line detection interrupt
-  *         - USRT1_IT_ERR:  Error interrupt
-  * @param  NewState new state of the specified USART interrupts.
-  *         This parameter can be: ENABLE or DISABLE.
-  * @retval None
-  */
-void UART1_ITConfig(UART1_IT_TypeDef UART1_IT, FunctionalState NewState)
-{
-  uint8_t uartreg = 0, itpos = 0x00;
-  
-  /* Check the parameters */
-  assert_param(IS_UART1_CONFIG_IT_OK(UART1_IT));
-  assert_param(IS_FUNCTIONALSTATE_OK(NewState));
-  
-  /* Get the UART1 register index */
-  uartreg = (uint8_t)((uint16_t)UART1_IT >> 0x08);
-  /* Get the UART1 IT index */
-  itpos = (uint8_t)((uint8_t)1 << (uint8_t)((uint8_t)UART1_IT & (uint8_t)0x0F));
-  
-  if (NewState != DISABLE)
-  {
-    /**< Enable the Interrupt bits according to UART1_IT mask */
-    if (uartreg == 0x01)
-    {
-      UART1->CR1 |= itpos;
-    }
-    else if (uartreg == 0x02)
-    {
-      UART1->CR2 |= itpos;
-    }
-    else
-    {
-      UART1->CR4 |= itpos;
-    }
-  }
-  else
-  {
-    /**< Disable the interrupt bits according to UART1_IT mask */
-    if (uartreg == 0x01)
-    {
-      UART1->CR1 &= (uint8_t)(~itpos);
-    }
-    else if (uartreg == 0x02)
-    {
-      UART1->CR2 &= (uint8_t)(~itpos);
-    }
-    else
-    {
-      UART1->CR4 &= (uint8_t)(~itpos);
-    }
-  }
-  
 }
 
 /**
@@ -265,7 +79,7 @@ void UART1_ITConfig(UART1_IT_TypeDef UART1_IT, FunctionalState NewState)
 void UART1_HalfDuplexCmd(FunctionalState NewState)
 {
   assert_param(IS_FUNCTIONALSTATE_OK(NewState));
-  
+
   if (NewState != DISABLE)
   {
     UART1->CR5 |= UART1_CR5_HDSEL;  /**< UART1 Half Duplex Enable  */
@@ -285,7 +99,7 @@ void UART1_HalfDuplexCmd(FunctionalState NewState)
 void UART1_IrDAConfig(UART1_IrDAMode_TypeDef UART1_IrDAMode)
 {
   assert_param(IS_UART1_IRDAMODE_OK(UART1_IrDAMode));
-  
+
   if (UART1_IrDAMode != UART1_IRDAMODE_NORMAL)
   {
     UART1->CR5 |= UART1_CR5_IRLP;
@@ -306,7 +120,7 @@ void UART1_IrDACmd(FunctionalState NewState)
 {
   /* Check parameters */
   assert_param(IS_FUNCTIONALSTATE_OK(NewState));
-  
+
   if (NewState != DISABLE)
   {
     /* Enable the IrDA mode by setting the IREN bit in the CR3 register */
@@ -329,7 +143,7 @@ void UART1_IrDACmd(FunctionalState NewState)
 void UART1_LINBreakDetectionConfig(UART1_LINBreakDetectionLength_TypeDef UART1_LINBreakDetectionLength)
 {
   assert_param(IS_UART1_LINBREAKDETECTIONLENGTH_OK(UART1_LINBreakDetectionLength));
-  
+
   if (UART1_LINBreakDetectionLength != UART1_LINBREAKDETECTIONLENGTH_10BITS)
   {
     UART1->CR4 |= UART1_CR4_LBDL;
@@ -349,7 +163,7 @@ void UART1_LINBreakDetectionConfig(UART1_LINBreakDetectionLength_TypeDef UART1_L
 void UART1_LINCmd(FunctionalState NewState)
 {
   assert_param(IS_FUNCTIONALSTATE_OK(NewState));
-  
+
   if (NewState != DISABLE)
   {
     /* Enable the LIN mode by setting the LINE bit in the CR2 register */
@@ -371,7 +185,7 @@ void UART1_LINCmd(FunctionalState NewState)
 void UART1_SmartCardCmd(FunctionalState NewState)
 {
   assert_param(IS_FUNCTIONALSTATE_OK(NewState));
-  
+
   if (NewState != DISABLE)
   {
     /* Enable the SC mode by setting the SCEN bit in the CR5 register */
@@ -394,7 +208,7 @@ void UART1_SmartCardCmd(FunctionalState NewState)
 void UART1_SmartCardNACKCmd(FunctionalState NewState)
 {
   assert_param(IS_FUNCTIONALSTATE_OK(NewState));
-  
+
   if (NewState != DISABLE)
   {
     /* Enable the NACK transmission by setting the NACK bit in the CR5 register */
@@ -416,7 +230,7 @@ void UART1_SmartCardNACKCmd(FunctionalState NewState)
 void UART1_WakeUpConfig(UART1_WakeUp_TypeDef UART1_WakeUp)
 {
   assert_param(IS_UART1_WAKEUP_OK(UART1_WakeUp));
-  
+
   UART1->CR1 &= ((uint8_t)~UART1_CR1_WAKE);
   UART1->CR1 |= (uint8_t)UART1_WakeUp;
 }
@@ -430,7 +244,7 @@ void UART1_WakeUpConfig(UART1_WakeUp_TypeDef UART1_WakeUp)
 void UART1_ReceiverWakeUpCmd(FunctionalState NewState)
 {
   assert_param(IS_FUNCTIONALSTATE_OK(NewState));
-  
+
   if (NewState != DISABLE)
   {
     /* Enable the mute mode UART1 by setting the RWU bit in the CR2 register */
@@ -448,20 +262,10 @@ void UART1_ReceiverWakeUpCmd(FunctionalState NewState)
   * @param  None
   * @retval The received data.
   */
-uint8_t UART1_ReceiveData8(void)
-{
-  return ((uint8_t)UART1->DR);
-}
-
-/**
-  * @brief  Returns the most recent received data by the UART1 peripheral.
-  * @param  None
-  * @retval The received data.
-  */
 uint16_t UART1_ReceiveData9(void)
 {
   uint16_t temp = 0;
-  
+
   temp = (uint16_t)(((uint16_t)( (uint16_t)UART1->CR1 & (uint16_t)UART1_CR1_R8)) << 1);
   return (uint16_t)( (((uint16_t) UART1->DR) | temp ) & ((uint16_t)0x01FF));
 }
@@ -512,7 +316,7 @@ void UART1_SetAddress(uint8_t UART1_Address)
 {
   /*assert_param for UART1_Address*/
   assert_param(IS_UART1_ADDRESS_OK(UART1_Address));
-  
+
   /* Clear the UART1 address */
   UART1->CR4 &= ((uint8_t)~UART1_CR4_ADD);
   /* Set the UART1 address node */
@@ -568,11 +372,11 @@ void UART1_SetPrescaler(uint8_t UART1_Prescaler)
 FlagStatus UART1_GetFlagStatus(UART1_Flag_TypeDef UART1_FLAG)
 {
   FlagStatus status = RESET;
-  
+
   /* Check parameters */
   assert_param(IS_UART1_FLAG_OK(UART1_FLAG));
-  
-  
+
+
   /* Check the status of the specified UART1 flag*/
   if (UART1_FLAG == UART1_FLAG_LBDF)
   {
@@ -624,29 +428,29 @@ FlagStatus UART1_GetFlagStatus(UART1_Flag_TypeDef UART1_FLAG)
   *         - UART1_FLAG_LBDF: LIN Break detection flag.
   *         - UART1_FLAG_RXNE: Receive data register not empty flag.
   * @note
-  *         - PE (Parity error), FE (Framing error), NE (Noise error), 
-  *         OR (OverRun error) and IDLE (Idle line detected) flags are 
+  *         - PE (Parity error), FE (Framing error), NE (Noise error),
+  *         OR (OverRun error) and IDLE (Idle line detected) flags are
   *         cleared by software sequence: a read operation to UART1_SR register
-  *         (UART1_GetFlagStatus())followed by a read operation to UART1_DR 
+  *         (UART1_GetFlagStatus())followed by a read operation to UART1_DR
   *         register(UART1_ReceiveData8() or UART1_ReceiveData9()).
-  *           
+  *
   *         - RXNE flag can be also cleared by a read to the UART1_DR register
   *         (UART1_ReceiveData8()or UART1_ReceiveData9()).
-  *           
+  *
   *         - TC flag can be also cleared by software sequence: a read operation
-  *         to UART1_SR register (UART1_GetFlagStatus()) followed by a write 
+  *         to UART1_SR register (UART1_GetFlagStatus()) followed by a write
   *         operation to UART1_DR register (UART1_SendData8() or UART1_SendData9()).
-  *           
-  *         - TXE flag is cleared only by a write to the UART1_DR register 
+  *
+  *         - TXE flag is cleared only by a write to the UART1_DR register
   *         (UART1_SendData8() or UART1_SendData9()).
-  *           
+  *
   *         - SBK flag is cleared during the stop bit of break.
   * @retval None
   */
 void UART1_ClearFlag(UART1_Flag_TypeDef UART1_FLAG)
 {
   assert_param(IS_UART1_CLEAR_FLAG_OK(UART1_FLAG));
-  
+
   /* Clear the Receive Register Not Empty flag */
   if (UART1_FLAG == UART1_FLAG_RXNE)
   {
@@ -679,25 +483,25 @@ ITStatus UART1_GetITStatus(UART1_IT_TypeDef UART1_IT)
   uint8_t itmask1 = 0;
   uint8_t itmask2 = 0;
   uint8_t enablestatus = 0;
-  
+
   /* Check parameters */
   assert_param(IS_UART1_GET_IT_OK(UART1_IT));
-  
+
   /* Get the UART1 IT index */
   itpos = (uint8_t)((uint8_t)1 << (uint8_t)((uint8_t)UART1_IT & (uint8_t)0x0F));
   /* Get the UART1 IT index */
   itmask1 = (uint8_t)((uint8_t)UART1_IT >> (uint8_t)4);
   /* Set the IT mask*/
   itmask2 = (uint8_t)((uint8_t)1 << itmask1);
-  
-  
+
+
   /* Check the status of the specified UART1 pending bit*/
   if (UART1_IT == UART1_IT_PE)
   {
     /* Get the UART1_IT enable bit status*/
     enablestatus = (uint8_t)((uint8_t)UART1->CR1 & itmask2);
     /* Check the status of the specified UART1 interrupt*/
-    
+
     if (((UART1->SR & itpos) != (uint8_t)0x00) && enablestatus)
     {
       /* Interrupt occurred*/
@@ -709,7 +513,7 @@ ITStatus UART1_GetITStatus(UART1_IT_TypeDef UART1_IT)
       pendingbitstatus = RESET;
     }
   }
-  
+
   else if (UART1_IT == UART1_IT_LBDF)
   {
     /* Get the UART1_IT enable bit status*/
@@ -742,7 +546,7 @@ ITStatus UART1_GetITStatus(UART1_IT_TypeDef UART1_IT)
       pendingbitstatus = RESET;
     }
   }
-  
+
   /* Return the UART1_IT status*/
   return  pendingbitstatus;
 }
@@ -754,20 +558,20 @@ ITStatus UART1_GetITStatus(UART1_IT_TypeDef UART1_IT)
   *         - UART1_IT_LBDF:  LIN Break detection interrupt
   *         - UART1_IT_RXNE: Receive Data register not empty interrupt.
   * @note
-  *         - PE (Parity error), FE (Framing error), NE (Noise error), 
-  *           OR (OverRun error) and IDLE (Idle line detected) pending bits are 
+  *         - PE (Parity error), FE (Framing error), NE (Noise error),
+  *           OR (OverRun error) and IDLE (Idle line detected) pending bits are
   *           cleared by software sequence: a read operation to UART1_SR register
   *           (UART1_GetITStatus()) followed by a read operation to UART1_DR register
   *           (UART1_ReceiveData8() or UART1_ReceiveData9()).
-  *   
+  *
   *         - RXNE pending bit can be also cleared by a read to the UART1_DR register
   *           (UART1_ReceiveData8() or UART1_ReceiveData9()).
-  * 
-  *         - TC (Transmit complete) pending bit can be cleared by software 
+  *
+  *         - TC (Transmit complete) pending bit can be cleared by software
   *           sequence: a read operation to UART1_SR register (UART1_GetITStatus())
   *           followed by a write operation to UART1_DR register (UART1_SendData8()
   *           or UART1_SendData9()).
-  *             
+  *
   *         - TXE pending bit is cleared only by a write to the UART1_DR register
   *           (UART1_SendData8() or UART1_SendData9()).
   * @retval None
@@ -775,7 +579,7 @@ ITStatus UART1_GetITStatus(UART1_IT_TypeDef UART1_IT)
 void UART1_ClearITPendingBit(UART1_IT_TypeDef UART1_IT)
 {
   assert_param(IS_UART1_CLEAR_IT_OK(UART1_IT));
-  
+
   /* Clear the Receive Register Not Empty pending bit */
   if (UART1_IT == UART1_IT_RXNE)
   {
@@ -791,10 +595,10 @@ void UART1_ClearITPendingBit(UART1_IT_TypeDef UART1_IT)
 /**
   * @}
   */
-  
+
 /**
   * @}
   */
-  
+
 
 /************************ (C) COPYRIGHT STMicroelectronics *****END OF FILE****/
