+++ src/stm8s_uart1_Init.c
@@ -1,0 +1,162 @@
+/**
+  ********************************************************************************
+  * @file    stm8s_uart1.c
+  * @author  MCD Application Team
+  * @version V2.2.0
+  * @date    30-September-2014
+  * @brief   This file contains all the functions for the UART1 peripheral.
+   ******************************************************************************
+  * @attention
+  *
+  * <h2><center>&copy; COPYRIGHT 2014 STMicroelectronics</center></h2>
+  *
+  * Licensed under MCD-ST Liberty SW License Agreement V2, (the "License");
+  * You may not use this file except in compliance with the License.
+  * You may obtain a copy of the License at:
+  *
+  *        http://www.st.com/software_license_agreement_liberty_v2
+  *
+  * Unless required by applicable law or agreed to in writing, software
+  * distributed under the License is distributed on an "AS IS" BASIS,
+  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  * See the License for the specific language governing permissions and
+  * limitations under the License.
+  *
+  ******************************************************************************
+  */
+
+/* Includes ------------------------------------------------------------------*/
+#include "stm8s_uart1.h"
+
+/** @addtogroup STM8S_StdPeriph_Driver
+  * @{
+  */
+
+/* Private typedef -----------------------------------------------------------*/
+/* Private define ------------------------------------------------------------*/
+/* Private macro -------------------------------------------------------------*/
+/* Private variables ---------------------------------------------------------*/
+/* Private function prototypes -----------------------------------------------*/
+/* Private functions ---------------------------------------------------------*/
+/* Public functions ----------------------------------------------------------*/
+
+/** @}
+  * @addtogroup UART1_Public_Functions
+  * @{
+  */
+
+/**
+  * @brief  Initializes the UART1 according to the specified parameters.
+  * @note   Configure in Push Pull or Open Drain mode the Tx pin by setting the
+  *         correct I/O Port register according the product package and line
+  *         configuration
+  * @param  BaudRate: The baudrate.
+  * @param  WordLength : This parameter can be any of the
+  *         @ref UART1_WordLength_TypeDef enumeration.
+  * @param  StopBits: This parameter can be any of the
+  *         @ref UART1_StopBits_TypeDef enumeration.
+  * @param  Parity: This parameter can be any of the
+  *         @ref UART1_Parity_TypeDef enumeration.
+  * @param  SyncMode: This parameter can be any of the
+  *         @ref UART1_SyncMode_TypeDef values.
+  * @param  Mode: This parameter can be any of the @ref UART1_Mode_TypeDef values
+  * @retval None
+  */
+void UART1_Init(uint32_t BaudRate, UART1_WordLength_TypeDef WordLength,
+                UART1_StopBits_TypeDef StopBits, UART1_Parity_TypeDef Parity,
+                UART1_SyncMode_TypeDef SyncMode, UART1_Mode_TypeDef Mode)
+{
+  uint32_t BaudRate_Mantissa = 0, BaudRate_Mantissa100 = 0;
+
+  /* Check the parameters */
+  assert_param(IS_UART1_BAUDRATE_OK(BaudRate));
+  assert_param(IS_UART1_WORDLENGTH_OK(WordLength));
+  assert_param(IS_UART1_STOPBITS_OK(StopBits));
+  assert_param(IS_UART1_PARITY_OK(Parity));
+  assert_param(IS_UART1_MODE_OK((uint8_t)Mode));
+  assert_param(IS_UART1_SYNCMODE_OK((uint8_t)SyncMode));
+
+  /* Clear the word length bit */
+  UART1->CR1 &= (uint8_t)(~UART1_CR1_M);
+
+  /* Set the word length bit according to UART1_WordLength value */
+  UART1->CR1 |= (uint8_t)WordLength;
+
+  /* Clear the STOP bits */
+  UART1->CR3 &= (uint8_t)(~UART1_CR3_STOP);
+  /* Set the STOP bits number according to UART1_StopBits value  */
+  UART1->CR3 |= (uint8_t)StopBits;
+
+  /* Clear the Parity Control bit */
+  UART1->CR1 &= (uint8_t)(~(UART1_CR1_PCEN | UART1_CR1_PS  ));
+  /* Set the Parity Control bit to UART1_Parity value */
+  UART1->CR1 |= (uint8_t)Parity;
+
+  /* Clear the LSB mantissa of UART1DIV  */
+  UART1->BRR1 &= (uint8_t)(~UART1_BRR1_DIVM);
+  /* Clear the MSB mantissa of UART1DIV  */
+  UART1->BRR2 &= (uint8_t)(~UART1_BRR2_DIVM);
+  /* Clear the Fraction bits of UART1DIV */
+  UART1->BRR2 &= (uint8_t)(~UART1_BRR2_DIVF);
+
+  /* Set the UART1 BaudRates in BRR1 and BRR2 registers according to UART1_BaudRate value */
+  BaudRate_Mantissa    = ((uint32_t)CLK_GetClockFreq() / (BaudRate << 4));
+  BaudRate_Mantissa100 = (((uint32_t)CLK_GetClockFreq() * 100) / (BaudRate << 4));
+  /* Set the fraction of UART1DIV  */
+  UART1->BRR2 |= (uint8_t)((uint8_t)(((BaudRate_Mantissa100 - (BaudRate_Mantissa * 100)) << 4) / 100) & (uint8_t)0x0F);
+  /* Set the MSB mantissa of UART1DIV  */
+  UART1->BRR2 |= (uint8_t)((BaudRate_Mantissa >> 4) & (uint8_t)0xF0);
+  /* Set the LSB mantissa of UART1DIV  */
+  UART1->BRR1 |= (uint8_t)BaudRate_Mantissa;
+
+  /* Disable the Transmitter and Receiver before setting the LBCL, CPOL and CPHA bits */
+  UART1->CR2 &= (uint8_t)~(UART1_CR2_TEN | UART1_CR2_REN);
+  /* Clear the Clock Polarity, lock Phase, Last Bit Clock pulse */
+  UART1->CR3 &= (uint8_t)~(UART1_CR3_CPOL | UART1_CR3_CPHA | UART1_CR3_LBCL);
+  /* Set the Clock Polarity, lock Phase, Last Bit Clock pulse */
+  UART1->CR3 |= (uint8_t)((uint8_t)SyncMode & (uint8_t)(UART1_CR3_CPOL |
+                                                        UART1_CR3_CPHA | UART1_CR3_LBCL));
+
+  if ((uint8_t)(Mode & UART1_MODE_TX_ENABLE))
+  {
+    /* Set the Transmitter Enable bit */
+    UART1->CR2 |= (uint8_t)UART1_CR2_TEN;
+  }
+  else
+  {
+    /* Clear the Transmitter Disable bit */
+    UART1->CR2 &= (uint8_t)(~UART1_CR2_TEN);
+  }
+  if ((uint8_t)(Mode & UART1_MODE_RX_ENABLE))
+  {
+    /* Set the Receiver Enable bit */
+    UART1->CR2 |= (uint8_t)UART1_CR2_REN;
+  }
+  else
+  {
+    /* Clear the Receiver Disable bit */
+    UART1->CR2 &= (uint8_t)(~UART1_CR2_REN);
+  }
+  /* Set the Clock Enable bit, lock Polarity, lock Phase and Last Bit Clock
+  pulse bits according to UART1_Mode value */
+  if ((uint8_t)(SyncMode & UART1_SYNCMODE_CLOCK_DISABLE))
+  {
+    /* Clear the Clock Enable bit */
+    UART1->CR3 &= (uint8_t)(~UART1_CR3_CKEN);
+  }
+  else
+  {
+    UART1->CR3 |= (uint8_t)((uint8_t)SyncMode & UART1_CR3_CKEN);
+  }
+}
+
+/**
+  * @}
+  */
+
+/**
+  * @}
+  */
+
+
+/************************ (C) COPYRIGHT STMicroelectronics *****END OF FILE****/
